### 2023-08-01 #5724 @editorialbot
[PRE REVIEW]: gcamreport: An R tool to process and standardize GCAM outputs
**Submitting author:** @klau506 (<a href="http://orcid.org/0000-0001-6696-7685">Clàudia  Rodés-Bachs</a>)
**Repository:** https://github.com/bc3LC/gcamreport
**Branch with paper.md** (empty if default branch): paper
**Version:** v6.0.1
