### 2023-08-04 #5724 @kthyng
Same question to @klau506: Is there a reason your comment is almost identical to on this other pre-review issue? https://github.com/openjournals/joss-reviews/issues/5722#issuecomment-1660820660
