### 2023-08-04 #5724 @klau506
@kthyng Thanks for your comments and feedback! The statement of need section already exists in the paper. If it should have a different format, please let me know. Regarding the suggestion of reviewers, it is similar to the one made in #5722 since both are packages to enlarge the functionalities of the same integrated assesment model; thus, looking for reviewers with similar backgrounds. If I can help you with anything else, do not hesitate to contact me!! :)
