### 2023-08-21 #5724 @klau506
@kthyng I have checked the syntax of the paper and did not find any issue that could prevent the "statement of need" section to be recognized. It complied perfectly with the Whedon webpage and the JOSS GitHub action, now integrated into the "paper" branch. If I can help with anything else, let me know! :)
