### 2023-09-18 #5724 @kthyng
Hi @klau506! We're just waiting on an editor — we've been inching up the waitlist and the paper before yours has just been taken on so I hope it won't be long now for yours. Thanks for your patience.
