### 2023-10-09 #5724 @abigailsnyder
@martinfleis COI disclosure: I work with one of the co-authors of this work (Zarrar Khan) at a 60ish person institution and we have been co-authors on a paper in the last year. Given that this institution is the home of GCAM, I think at least one reviewer from this institution reviewing this paper is probably desirable, but I leave it to your discretion whether I am the right person.
Another of the coauthors (Ryna Cui) and I have also been coauthors on some large-author-list papers, but not in the last 3 years at least I think.

Please let me know if these potential COIs are an issue
