### 2023-10-16 #5724 @abigailsnyder
@martinfleis  I only did a cursory check of recent publications for the following, and I'm not sure any of them have bandwidth but this list might be a good start:

[Ben Bond-Lamberty](https://github.com/bpbond)
[Kalyn R. Dorheim](https://github.com/kdorheim)
[marideeweber](https://github.com/marideeweber)
[Noah S. Prime](https://github.com/noahprime)
[Kendalynn Morris](https://github.com/kendalynnm)
[rhoesly](https://github.com/rhoesly)
