### 2023-10-23 #5975 @editorialbot
[REVIEW]: gcamreport: An R tool to process and standardize GCAM outputs
**Submitting author:** @klau506 (<a href="http://orcid.org/0000-0001-6696-7685">Clàudia  Rodés-Bachs</a>)
**Repository:** https://github.com/bc3LC/gcamreport
**Branch with paper.md** (empty if default branch): gcam-v7.0
**Version:** v7.0.1
