### 2023-10-25 b27973d
Update queries gcam7. Optimize data_queries function
