### 2023-10-30 ea82f1f
Fix data_query function
