+++ inst/extdata/saveDataFiles.R
@@ -7,8 +7,8 @@ library(here)
 rawDataFolder = here::here()
 
 # emissions considered nonCO2
-emissions_list = c('BC','BC_AWB','C2F6','CF4','CH4','CH4_AGR','CH4_AWB','CO','CO2','CO_AWB','H2',
-                   'H2_AWB','HFC125','HFC134a','HFC143a','HFC152a','HFC227ea','HFC23','HFC236fa',
+emissions_list = c('BC','BC_AWB','C2F6','CF4','CH4','CH4_AGR','CH4_AWB','CO','CO2','CO_AWB','CO2_ETS',
+                   'H2','H2_AWB','HFC125','HFC134a','HFC143a','HFC152a','HFC227ea','HFC23','HFC236fa',
                    'HFC245fa','HFC32','HFC365mfc','HFC43','N2O','N2O_AGR','N2O_AWB','NH3','NH3_AGR',
                    'NH3_AWB','NMVOC','NMVOC_AGR','NMVOC_AWB','NOx','NOx_AGR','NOx_AWB','OC','OC_AWB',
                    'PM10','PM2.5','SF6','SO2_1','SO2_1_AWB','SO2_2','SO2_2_AWB','SO2_3','SO2_3_AWB',
@@ -47,6 +47,10 @@ use_data(template, overwrite=T)
 co2_sector_map <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/CO2_sector_map.csv"), skip = 1, na = "",
                            stringsAsFactors = FALSE) %>% gather_map()
 use_data(co2_sector_map, overwrite=T)
+
+co2_ets_sector_map <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/CO2_ETS_sector_map.csv"), skip = 1, na = "",
+                           stringsAsFactors = FALSE) %>% gather_map()
+use_data(co2_ets_sector_map, overwrite=T)
 
 co2_tech_map <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/CO2_tech_map.csv"), skip = 1, na = "",
                          stringsAsFactors = FALSE) %>% gather_map()
@@ -154,6 +158,10 @@ investment <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/investm
   dplyr::mutate(value = gsub("%", "", value)) %>%
   dplyr::mutate(value = as.numeric(value))
 use_data(investment, overwrite=T)
+
+
+last_historical_year = 2015
+use_data(last_historical_year, overwrite=T)
 
 
 carbon_content <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/L202.CarbonCoef.csv"), skip = 2, na = "",
