+++ man/get_co2_price_share.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/functions.R
+\name{get_co2_price_share}
+\alias{get_co2_price_share}
+\title{get_co2_price_share}
+\usage{
+get_co2_price_share()
+}
+\value{
+co2_price_share global variable
+}
+\description{
+Get co2 price between CO2 and CO2_ETS. If only one CO2 type present, share = 1;
+otherwise, each type has the share corresponding to the last historical year
+}
+\keyword{co2}
+\keyword{rgcam::getQuery}
+\keyword{tmp}
