### 2023-11-01 e2560d1
Fix CO2_ETS factor conversion
