### 2023-11-02 0d19ce5
Generalize CO2_ETS
