+++ R/functions.R
@@ -6,6 +6,7 @@ options(dplyr.summarise.inform = FALSE)
 #                           ANCILLARY FUNCTIONS                         #
 #########################################################################
 
+
 #' conv_ghg_co2e
 #'
 #' Covert GHG to CO2e
@@ -251,15 +252,25 @@ get_co2 = function() {
 #'
 #' Get World's CO2 ETS emissions query.
 #' @keywords co2 rgcam::getQuery
-#' @return co2_ets global variable
+#' @return co2_ets_by reg and co2_ets_bysec global variables
 #' @export
 get_co2_ets = function() {
-  co2_ets <<-
+  co2_ets_byreg <<-
     tibble::as_tibble(rgcam::getQuery(prj, "nonCO2 emissions by region")) %>%
     dplyr::filter(ghg == 'CO2_ETS') %>%
     # change units to CO2 equivalent and set the variable
     dplyr::mutate(value = value * CO2_equivalent,
                   var = 'Emissions|CO2_ETS|Energy and Industrial Processes') %>%
+    dplyr::select(all_of(long_columns))
+  co2_ets_bysec <<-
+    tibble::as_tibble(rgcam::getQuery(prj, "nonCO2 emissions by sector (excluding resource production)")) %>%
+    dplyr::filter(ghg == 'CO2_ETS') %>%
+    # change units to CO2 equivalent and group by sector
+    dplyr::left_join(co2_ets_sector_map, by = "sector", multiple = "all") %>%
+    dplyr::mutate(value = value * unit_conv) %>%
+    dplyr::group_by(scenario, region, year, var) %>% #
+    dplyr::summarise(value = sum(value, na.rm = T)) %>%
+    dplyr::ungroup() %>%
     dplyr::select(all_of(long_columns))
 }
 
@@ -1004,22 +1015,51 @@ get_co2_price_global_tmp = function() {
 #' Get co2 price between CO2 and CO2_ETS. If only one CO2 type present, share = 1;
 #' otherwise, each type has the share corresponding to the last historical year
 #' @keywords co2 rgcam::getQuery tmp
-#' @return co2_price_share global variable
+#' @return co2_price_share_byreg and co2_price_share_bysec global variables
 #' @export
 get_co2_price_share = function() {
-  co2_price_share <<- co2_clean %>%
+  co2_price_share_byreg <<- co2_clean %>%
     dplyr::filter(var == 'Emissions|CO2|Energy and Industrial Processes',
                   year == last_historical_year) %>%
-    rbind(co2_ets %>%
+    rbind(co2_ets_byreg %>%
             dplyr::filter(var == 'Emissions|CO2_ETS|Energy and Industrial Processes',
                           year == last_historical_year)) %>%
-    dplyr::group_by(scenario, region, year) %>%
-    dplyr::mutate(share = value / sum(value)) %>%
-    dplyr::rename(market_adj = var) %>%
-    dplyr::mutate(market_adj = dplyr::if_else(market_adj == "Emissions|CO2|Energy and Industrial Processes", 'CO2', 'CO2_ETS')) %>%
-    dplyr::ungroup() %>%
-    # if CO2, share = 1
-    dplyr::mutate(share = dplyr::if_else(market_adj == 'CO2', 1, share))
+    dplyr::mutate(var = dplyr::if_else(var == "Emissions|CO2|Energy and Industrial Processes", 'CO2', 'CO2_ETS')) %>%
+    tidyr::pivot_wider(names_from = 'var', values_from = 'value') %>%
+    dplyr::mutate(CO2_ETS = dplyr::if_else(is.na(CO2_ETS), 0, CO2_ETS)) %>%
[…]
