+++ R/main.R
@@ -10,15 +10,25 @@ library(magrittr)
 #' @param prj_name: name of the project
 #' @param scenarios: name of the scenarios to be considered
 #' @param type: either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'
+#' @param desired_regions: desired regions to consider. By default, 'All' (written as NULL). Otherwise, specify a vector
+#' with all the considered regions, being USA,Africa_Eastern,Africa_Northern,Africa_Southern,Africa_Western,Australia_NZ,Brazil,Canada,
+#' Central Asia,China,EU-12,EU-15,Europe_Eastern,Europe_Non_EU,European Free Trade Association,India,Indonesia,Japan,
+#' Mexico,Middle East,Pakistan,Russia,South Africa,South America_Northern,South America_Southern,South Asia,South Korea,
+#' Southeast Asia,Taiwan,Argentina,Colombia,Central America and Caribbean. ATTENCION: the considered regions will make up "World".
+#' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @return dataframe with the data from the query
 #' @export
-data_query = function(type, db_path, db_name, prj_name, scenarios) {
+data_query = function(type, db_path, db_name, prj_name, scenarios, desired_regions = 'All') {
+  if (length(desired_regions) == 1 && desired_regions == 'All') {
+    desired_regions = NULL
+  }
+
   dt = data.frame()
   xml <- xml2::read_xml('inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml')
   qq <- xml2::xml_find_first(xml, paste0("//*[@title='", type, "']"))
 
-  emiss_list = emissions_list
   for (sc in scenarios) {
+    emiss_list = emissions_list
     while (length(emiss_list) > 0) {
       current_emis = emiss_list[1:min(21,length(emiss_list))]
       qq_sec = gsub("current_emis", paste0("(@name = '", paste(current_emis, collapse = "' or @name = '"), "')"), qq)
@@ -30,7 +40,7 @@ data_query = function(type, db_path, db_name, prj_name, scenarios) {
         qn = type,
         query = qq_sec,
         scenario = sc,
-        regions = NULL,
+        regions = desired_regions,
         clobber = TRUE,
         transformations = NULL,
         saveProj = FALSE,
@@ -65,13 +75,19 @@ data_query = function(type, db_path, db_name, prj_name, scenarios) {
 #' @param db_name: name of the database
 #' @param prj_name: name of the project
 #' @param scenarios: name of the scenarios to be considered
+#' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector
+#' with all the considered regions, being USA,Africa_Eastern,Africa_Northern,Africa_Southern,Africa_Western,Australia_NZ,Brazil,Canada,
+#' Central Asia,China,EU-12,EU-15,Europe_Eastern,Europe_Non_EU,European Free Trade Association,India,Indonesia,Japan,
+#' Mexico,Middle East,Pakistan,Russia,South Africa,South America_Northern,South America_Southern,South Asia,South Korea,
+#' Southeast Asia,Taiwan,Argentina,Colombia,Central America and Caribbean. ATTENCION: the considered regions will make up "World".
+#' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @return empty dataframes on the void queries of the project
 #' @export
-fill_queries = function(db_path, db_name, prj_name, scenarios) {
+fill_queries = function(db_path, db_name, prj_name, scenarios, desired_regions = 'All') {
   # add nonCO2 queries manually (they are too big to use the usual method)
[…]
