+++ R/functions.R
@@ -6,6 +6,45 @@ options(dplyr.summarise.inform = FALSE)
 #                           ANCILLARY FUNCTIONS                         #
 #########################################################################
 
+
+#' filter_regions
+#'
+#' Filter the desired regions of a GCAM project
+#' @param data: dataframe to be filtered
+#' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions,
+#' being USA,Africa_Eastern,Africa_Northern,Africa_Southern,Africa_Western,Australia_NZ,Brazil,Canada,
+#' Central Asia,China,EU-12,EU-15,Europe_Eastern,Europe_Non_EU,European Free Trade Association,India,Indonesia,Japan,
+#' Mexico,Middle East,Pakistan,Russia,South Africa,South America_Northern,South America_Southern,South Asia,South Korea,
+#' Southeast Asia,Taiwan,Argentina,Colombia,Central America and Caribbean. ATTENCION: the considered regions will make up "World".
+#' In case the project dataset needs to be created, it will be produced with only the specified regions
+#' @param variable: dataset variable information
+#' @return filtered dataframe
+#' @export
+filter_regions <- function (data, desired_regions = 'All', variable) {
+  if (!(length(desired_regions) == 1 && desired_regions == 'All')) {
+    # the variable CO2 prices does not contain "region", but "markets". Now we
+    # filter for all market items that do not contain the desired regions
+    if (variable %in% c('CO2 prices','supply of all markets')) {
+      pattern <- paste(c("CO2", "airCO2", "nonCO2", "CO2_FUG", "CO2 removal",
+                         "H2", "Exports"), collapse = "|")
+      data = data %>%
+        dplyr::mutate(region = sapply(strsplit(as.character(market), pattern),
+                                      function(x) x[1])) %>%
+        dplyr::filter(region %in% desired_regions) %>%
+        dplyr::select(-region)
+    } else if (!(v %in% c('CO2 concentrations','global mean temperature',
+                          'total climate forcing'))) {
+      data = data %>%
+        dplyr::filter(region %in% desired_regions)
+    }
+  }
+
+  return(data)
+}
+
+
+
+
 #' conv_ghg_co2e
 #'
 #' Covert GHG to CO2e
@@ -918,6 +957,23 @@ get_price_var_tmp = function() {
 }
 
 
+
+#' filter_data_regions
+#'
+#' Filter the desired regions of some data with "regions" column
+#' @keywords internal tmp process
+#' @return data containing only the desired regions
+#' @export
+filter_data_regions = function(data) {
+  if (!(length(desired_regions) == 1 && desired_regions == 'All')) {
+    data = data %>%
+      dplyr::filter(region %in% desired_regions)
+  }
+
+  return(data)
+}
+
+
 #' get_regions_tmp
 #'
 #' Get regions to compute carbon price.
@@ -925,8 +981,9 @@ get_price_var_tmp = function() {
 #' @return regions global variable
 #' @export
 get_regions_tmp = function() {
+  CO2_market_filteredReg = filter_data_regions(CO2_market)
   regions <<-
-    unique(CO2_market$region)
+    unique(CO2_market_filteredReg$region)
 }
 
 
@@ -991,26 +1048,30 @@ get_co2_price_fragmented_tmp = function() {
   co2_price_fragmented_pre <<-
     rgcam::getQuery(prj, "CO2 prices") %>%
[…]
