### 2023-11-28 e00bcf3
Integrate regions selection into IAM COMPACT branch
