+++ man/filter_regions.Rd
@@ -1,30 +1,26 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/main.R
-\name{create_project}
-\alias{create_project}
-\title{create_project}
+% Please edit documentation in R/functions.R
+\name{filter_regions}
+\alias{filter_regions}
+\title{filter_regions}
 \usage{
-create_project(db_path, db_name, prj_name, scenarios, desired_regions = "All")
+filter_regions(data, desired_regions = "All", variable)
 }
 \arguments{
-\item{db_path:}{path of the database}
-
-\item{db_name:}{name of the database}
-
-\item{prj_name:}{name of the project}
-
-\item{scenarios:}{name of the scenarios to be considered}
+\item{data:}{dataframe to be filtered}
 
 \item{desired_regions:}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions,
 being USA,Africa_Eastern,Africa_Northern,Africa_Southern,Africa_Western,Australia_NZ,Brazil,Canada,
 Central Asia,China,EU-12,EU-15,Europe_Eastern,Europe_Non_EU,European Free Trade Association,India,Indonesia,Japan,
 Mexico,Middle East,Pakistan,Russia,South Africa,South America_Northern,South America_Southern,South Asia,South Korea,
 Southeast Asia,Taiwan,Argentina,Colombia,Central America and Caribbean. ATTENCION: the considered regions will make up "World".
-In case the project dataset needs to be created, it will be produced with only the specified regions.}
+In case the project dataset needs to be created, it will be produced with only the specified regions}
+
+\item{variable:}{dataset variable information}
 }
 \value{
-loaded project into global environment
+filtered dataframe
 }
 \description{
-Create specified project and load it into the global environment
+Filter the desired regions of a GCAM project
 }
