+++ NAMESPACE
@@ -19,6 +19,8 @@ export(do_data_sample)
 export(do_mount_tree)
 export(do_unmount_tree)
 export(fill_queries)
+export(filter_data_regions)
+export(filter_regions)
 export(firstup)
 export(gather_map)
 export(get_CCS_invest)

[…]
