### 2023-11-30 46d7e63
Fix selection of variables
