### 2023-11-30 e3af0f7
Fix selection of variables
