### 2023-11-30 267a35f
use same capacity factors for additions and total
