### 2023-12-01 b0b8933
Add avail_reg, var, cont functions and final reporting variables dependency with internal varialbes
