+++ R/functions.R
@@ -7,20 +7,16 @@ options(dplyr.summarise.inform = FALSE)
 #########################################################################
 
 
-#' filter_regions
+#' filter_loading_regions
 #'
 #' Filter the desired regions of a GCAM project
 #' @param data: dataframe to be filtered
-#' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions,
-#' being USA,Africa_Eastern,Africa_Northern,Africa_Southern,Africa_Western,Australia_NZ,Brazil,Canada,
-#' Central Asia,China,EU-12,EU-15,Europe_Eastern,Europe_Non_EU,European Free Trade Association,India,Indonesia,Japan,
-#' Mexico,Middle East,Pakistan,Russia,South Africa,South America_Northern,South America_Southern,South Asia,South Korea,
-#' Southeast Asia,Taiwan,Argentina,Colombia,Central America and Caribbean. ATTENCION: the considered regions will make up "World".
-#' In case the project dataset needs to be created, it will be produced with only the specified regions
+#' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
+#' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
+#' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @param variable: dataset variable information
 #' @return filtered dataframe
-#' @export
-filter_regions <- function (data, desired_regions = 'All', variable) {
+filter_loading_regions <- function (data, desired_regions = 'All', variable) {
   if (!(length(desired_regions) == 1 && desired_regions == 'All')) {
     # the variable CO2 prices does not contain "region", but "markets". Now we
     # filter for all market items that do not contain the desired regions
@@ -51,6 +47,22 @@ filter_regions <- function (data, desired_regions = 'All', variable) {
 }
 
 
+#' filter_variables
+#'
+#' Filter the desired regions of a GCAM project
+#' @param data: dataframe to be filtered
+#' @return filtered dataframe
+#' @export
+filter_variables = function(data) {
+  if (!(length(desired_variables) == 1 & desired_variables == 'All')) {
+    if ('var' %in% colnames(data)) {
+      data = data %>%
+        dplyr::filter(var %in% desired_variables)
+    }
+  }
+
+  return(invisible(data))
+}
 
 
 #' conv_ghg_co2e
@@ -285,7 +297,7 @@ get_co2_concentration = function() {
 get_co2 = function() {
   co2_clean <<-
     tibble::as_tibble(rgcam::getQuery(prj, "CO2 emissions by sector (no bio) (excluding resource production)")) %>%
-    dplyr::left_join(co2_sector_map, by = "sector", multiple = "all") %>%
+    dplyr::left_join(filter_variables(co2_sector_map), by = "sector", multiple = "all") %>%
     dplyr::mutate(value = value * unit_conv) %>%
     dplyr::group_by(scenario, region, year, var) %>% #
     dplyr::summarise(value = sum(value, na.rm = T)) %>%
@@ -312,7 +324,7 @@ get_co2_ets = function() {
     tibble::as_tibble(rgcam::getQuery(prj, "nonCO2 emissions by sector (excluding resource production)")) %>%
     dplyr::filter(ghg == 'CO2_ETS') %>%
[…]
