+++ man/available_continents.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/main.R
+\name{available_continents}
+\alias{available_continents}
+\title{available_continents}
+\usage{
+available_continents(print = TRUE)
+}
+\arguments{
+\item{print:}{if TRUE, prints all available continents/regions' groups. TRUE by default.}
+}
+\value{
+Prints a list of all the available continents/regions' groups for the
+IAMC reporting dataset. It also returns them as a vector.
+}
+\description{
+available_continents
+}
