+++ man/update_template.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/functions.R
+\name{update_template}
+\alias{update_template}
+\title{update_template}
+\usage{
+update_template()
+}
+\value{
+Updated template as .rda and as csv in the inst/extdata folder
+}
+\description{
+Update the template file. This function checks for new reported variables,
+as well as cleans no-reported variables
+}
+\keyword{internal}
