+++ man/available_regions.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/main.R
+\name{available_regions}
+\alias{available_regions}
+\title{available_regions}
+\usage{
+available_regions(print = TRUE)
+}
+\arguments{
+\item{print:}{if TRUE, prints all available regions. TRUE by default.}
+}
+\value{
+Prints a list of all the available regions for the IAMC reporting dataset.
+It also returns them as a vector.
+}
+\description{
+available_regions
+}
