+++ man/available_variables.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/main.R
+\name{available_variables}
+\alias{available_variables}
+\title{available_variables}
+\usage{
+available_variables(print = TRUE)
+}
+\arguments{
+\item{print:}{if TRUE, prints all available variables. TRUE by default.}
+}
+\value{
+Prints a list of all the available variables for the IAMC reporting dataset.
+It also returns them as a vector.
+}
+\description{
+available_variables
+}

[…]
