### 2023-12-01 47380b9
Add avail_reg, var, cont functions and final reporting variables dependency with internal varialbes
