+++ man/filter_regions.Rd
@@ -1,26 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/functions.R
-\name{filter_regions}
-\alias{filter_regions}
-\title{filter_regions}
-\usage{
-filter_regions(data, desired_regions = "All", variable)
-}
-\arguments{
-\item{data:}{dataframe to be filtered}
-
-\item{desired_regions:}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions,
-being USA,Africa_Eastern,Africa_Northern,Africa_Southern,Africa_Western,Australia_NZ,Brazil,Canada,
-Central Asia,China,EU-12,EU-15,Europe_Eastern,Europe_Non_EU,European Free Trade Association,India,Indonesia,Japan,
-Mexico,Middle East,Pakistan,Russia,South Africa,South America_Northern,South America_Southern,South Asia,South Korea,
-Southeast Asia,Taiwan,Argentina,Colombia,Central America and Caribbean. ATTENCION: the considered regions will make up "World".
-In case the project dataset needs to be created, it will be produced with only the specified regions}
-
-\item{variable:}{dataset variable information}
-}
-\value{
-filtered dataframe
-}
-\description{
-Filter the desired regions of a GCAM project
-}
