+++ man/get_gov_revenue.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/functions.R
-\name{get_gov_revenue_all}
-\alias{get_gov_revenue_all}
-\title{get_gov_revenue_all}
+\name{get_gov_revenue}
+\alias{get_gov_revenue}
+\title{get_gov_revenue}
 \usage{
-get_gov_revenue_all()
+get_gov_revenue()
 }
 \value{
 gov_revenue_all global variable
