### 2023-12-01 fd650fd
Allow project creation with required queries for the desired_variables
