+++ man/load_query.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/main.R
+\name{load_query}
+\alias{load_query}
+\title{load_query}
+\usage{
+load_query(var, base_data, final_queries)
+}
+\arguments{
+\item{var:}{variable to be loaded}
+
+\item{base_data:}{dataframe with the required internal variables}
+
+\item{final_queries:}{vector of the queries to be loaded}
+}
+\value{
+query name to be loaded
+}
+\description{
+Recursive function to load the necessary queries for the desired variables
+}
+\keyword{internal}
