### 2023-12-01 ca7cf0e
Allow project creation with required queries for the desired_variables
