+++ R/main.R
@@ -76,11 +76,14 @@ data_query = function(type, db_path, db_name, prj_name, scenarios, desired_regio
 #' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @param prj: prject file
+#' @param desired_queries: desired queries to be loaded
 #' @return project file with extra queries
 #' @export
-fill_queries = function(db_path, db_name, prj_name, scenarios, desired_regions = 'All', prj) {
+fill_queries = function(db_path, db_name, prj_name, scenarios,
+                        desired_regions = 'All', prj, desired_queries) {
   # add nonCO2 queries manually (they are too big to use the usual method)
-  if (!'nonCO2 emissions by sector (excluding resource production)' %in% rgcam::listQueries(prj)) {
+  if (!'nonCO2 emissions by sector (excluding resource production)' %in% rgcam::listQueries(prj) &
+      'nonCO2 emissions by sector (excluding resource production)' %in% desired_queries) {
     print('nonCO2 emissions by sector (excluding resource production)')
     dt_sec = data_query('nonCO2 emissions by sector (excluding resource production)', db_path, db_name, prj_name, scenarios, desired_regions)
     prj_tmp <- rgcam::addQueryTable(project = 'prj_tmp1.dat', qdata = dt_sec, saveProj = FALSE,
@@ -88,7 +91,8 @@ fill_queries = function(db_path, db_name, prj_name, scenarios, desired_regions =
     prj <<- rgcam::mergeProjects(prj_name, list(prj,prj_tmp), clobber = TRUE, saveProj = FALSE)
 
   }
-  if (!'nonCO2 emissions by region' %in% rgcam::listQueries(prj)) {
+  if (!'nonCO2 emissions by region' %in% rgcam::listQueries(prj) &
+      'nonCO2 emissions by region' %in% desired_queries) {
     print('nonCO2 emissions by region')
     dt_reg = data_query('nonCO2 emissions by region', db_path, db_name, prj_name, scenarios, desired_regions)
     prj_tmp <- rgcam::addQueryTable(project = 'prj_tmp2.dat', qdata = dt_reg, saveProj = FALSE,
@@ -97,7 +101,8 @@ fill_queries = function(db_path, db_name, prj_name, scenarios, desired_regions =
   }
 
   # fix CO2 prices if needed
-  if (!'CO2 prices' %in% rgcam::listQueries(prj)) {
+  if (!'CO2 prices' %in% rgcam::listQueries(prj) &
+      'CO2 prices' %in% desired_queries) {
     l = length(rgcam::listScenarios(prj))
     dt = data.frame(Units = rep(NA,l),
                     scenario = rgcam::listScenarios(prj),
@@ -154,7 +159,9 @@ load_project = function(project_path, desired_regions = 'All') {
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @param desired_variables: desired variables to have in the report. Considered 'All' by default.
 #' Otherwise, specify a vector with all the desired options. To know all possible options, run `available_variables()`.
-#' In case the project dataset needs to be created, it will be produced with only the specified variables.
+#' In case the project dataset needs to be created, it will be produced with only the specified variables. ATTENTION:
+#' the global variables such as "Emissions" will be computed considering only the selected variables, for instance "Emissions|CO2",
+#' and will no account for other variables, such as "Emissions|CH4" or "Emissions|NH3".
 #' @return loaded project into global environment
 #' @export
 create_project = function(db_path, db_name, prj_name, scenarios,
@@ -165,16 +172,42 @@ create_project = function(db_path, db_name, prj_name, scenarios,
 
   # read the queries file
[…]
