+++ NAMESPACE
@@ -93,6 +93,7 @@ export(get_transmission_invest)
 export(is_leaf)
 export(launch_gcamreport_ui)
 export(load_project)
+export(load_query)
 export(load_variable)
 export(reset_first_load)
 export(run)
