### 2023-12-01 d3cb667
Check if project exists even if the project_path is not specified
