+++ R/main.R
@@ -128,6 +128,8 @@ fill_queries = function(db_path, db_name, prj_name, scenarios,
 #' @return loaded project into global environment
 #' @export
 load_project = function(project_path, desired_regions = 'All') {
+  print('Loading project...')
+
   # load the project
   prj = rgcam::loadProject(project_path)
 
@@ -166,77 +168,92 @@ load_project = function(project_path, desired_regions = 'All') {
 #' @export
 create_project = function(db_path, db_name, prj_name, scenarios,
                           desired_regions = 'All', desired_variables = 'All') {
-  # create the project
-  conn <- rgcam::localDBConn(db_path,
-                             db_name,migabble = FALSE)
 
-  # read the queries file
-  queryFile = paste0('inst/extdata/queries/','queries_gcamreport_gcam7.0_complete.xml')
-  queries_original <- rgcam::parse_batch_query(queryFile)
+  # check if the project already exists
+  file_name = paste0(db_path, "/", db_name, '_', prj_name)
+  if (file.exists(file_name)) {
+    print('Loading project...')
+    prj <<- rgcam::loadProject(file_name)
 
-  # subset the queries necessary for the selected variables
-  if (!(length(desired_variables) == 1 && desired_variables == 'All')) {
-    required_internal_variables = var_fun_map %>%
-      dplyr::rename('Internal_variable' = 'name') %>%
-      dplyr::left_join(template %>%
-                         dplyr::filter(!is.na(Internal_variable)),
-                       by = 'Internal_variable') %>%
-      dplyr::mutate(required = dplyr::if_else(Variable %in% desired_variables, TRUE, FALSE))
+  } else {
+    print('Creating project...')
 
-    required_queries = c()
-    for (req_int_var in unique(required_internal_variables %>%
-                     dplyr::filter(required == TRUE) %>%
-                     dplyr::pull(Variable))) {
-      required_queries = c(required_queries,
-                           load_query(required_internal_variables[which(required_internal_variables$Variable == req_int_var),],
-                                      required_internal_variables,
-                                      c()))
+    # create the project
+    conn <- rgcam::localDBConn(db_path,
+                               db_name,migabble = FALSE)
+
+    # read the queries file
+    queryFile = paste0('inst/extdata/queries/','queries_gcamreport_gcam7.0_complete.xml')
+    queries_original <- rgcam::parse_batch_query(queryFile)
+
+    # subset the queries necessary for the selected variables
+    if (!(length(desired_variables) == 1 && desired_variables == 'All')) {
+      # create a mapping with the Variables, Internal variables, functions to load
+      # them, and the dependencies
+      required_internal_variables = var_fun_map %>%
+        dplyr::rename('Internal_variable' = 'name') %>%
+        dplyr::left_join(template %>%
+                           dplyr::filter(!is.na(Internal_variable)),
+                         by = 'Internal_variable') %>%
+        dplyr::mutate(required = dplyr::if_else(Variable %in% desired_variables, TRUE, FALSE))
+
+      # create a vector with the required queries for the desired variables
+      required_queries = c()
+      for (req_int_var in unique(required_internal_variables %>%
+                       dplyr::filter(required == TRUE) %>%
+                       dplyr::pull(Variable))) {
+        required_queries = c(required_queries,
+                             load_query(required_internal_variables[which(required_internal_variables$Variable == req_int_var),],
+                                        required_internal_variables,
+                                        c()))
+      }
[…]
