### 2023-12-01 5202195
Improve vetting: print messages and allow summary check
