### 2023-12-01 2c0c098
Improve vetting: print messages and allow summary check
