### 2023-12-01 e053047
Flexibilize launch_ui function
