+++ R/main.R
@@ -380,10 +380,11 @@ available_variables = function(print = TRUE) {
 
 #' run
 #'
-#' Main function. Interacts with the user to select the desired variables for the report, loads
-#' them, saves them in an external output, runs the verifications, and informs the user about the
-#' success of the whole process. Either the `project_path` should be specified, or the `db_path`
-#' with all the related items, being `db_name`, `prj_name`, and `scenarios`.
+#' Main function. Creates/loads a GCAM project, standardizes the data and saves it
+#' in several forms (RData, CSV, and XLSX), runs vetting verifications, and launches
+#' the User Interface. Specify regions and/or variables to be reported and point
+#' either the `project_path`, or the `db_path`, the `db_name`, the `prj_name`, and `scenarios`.
+#' The resulting RData output can be used to call manually `launch_gcamreport_ui`.
 #' @param project_path: full path of the project with the project name. Possible extensions: .dat and .proj.
 #' @param db_path: full path of the database.
 #' @param db_name: name of the database.
@@ -392,9 +393,10 @@ available_variables = function(print = TRUE) {
 #' @param final_year: final year of the data. By default = 2100. ATENTION: final_year must be at least 2025.
 #' @param desired_variables: desired variables to have in the report. Considered 'All' by default.
 #' Otherwise, specify a vector with all the desired options. To know all possible options, run `available_variables()`.
-#' In case the project dataset needs to be created, it will be produced with only the specified variables. ATTENTION:
-#' the global variables such as "Emissions" will be computed considering only the selected variables, for instance "Emissions|CO2",
-#' and will no account for other variables, such as "Emissions|CH4" or "Emissions|NH3"
+#' In case the project dataset needs to be created, it will be created containing only the specified variables.
+#' ATTENTION: global variables such as "Emissions" will be computed considering only the selected variables. As an example,
+#' if you select "Emissions" and "Emissions|CO2", "Emissions" will only account for "Emissions|CO2", and will not
+#' consider other variables, such as "Emissions|CH4" or "Emissions|NH3"
 #' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
 #' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
@@ -402,16 +404,16 @@ available_variables = function(print = TRUE) {
 #' the considered continents/regions' groups. To know all possibilities, run `available_continents()`. ATTENTION: the considered
 #' continents/regions' groups will make up "World". In case the project dataset needs to be created, it will be produced with only the
 #' specified continents/regions' groups.
-#' @param save_output: if TRUE, save reporting data in CSV and XLSX formats. If FALSE, do not save data. If equals 'CSV' or 'XLSX',
-#' data saved only in the specified format.
-#' @param file_name: file path and name of the saved data. Not used if data not saved. By default, saved in the same directory and with
-#' the same name than the specified project_path, with 'iamc_report' tag. CSV and XLSX output. In case of specifing the path, do not
-#' introduce the extension, it will be automatically added.
-#' @param launch_ui: if TRUE, launch UI, Do not launch UI otherwise.
[…]
