### 2023-12-01 33e76c7
Flexibilize launch_ui function
