+++ man/fill_queries.Rd
@@ -1,39 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/main.R
-\name{fill_queries}
-\alias{fill_queries}
-\title{fill_queries}
-\usage{
-fill_queries(
-  db_path,
-  db_name,
-  prj_name,
-  scenarios,
-  desired_regions = "All",
-  prj,
-  desired_queries
-)
-}
-\arguments{
-\item{db_path:}{path of the database}
-
-\item{db_name:}{name of the database}
-
-\item{prj_name:}{name of the project}
-
-\item{scenarios:}{name of the scenarios to be considered}
-
-\item{desired_regions:}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
-To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
-In case the project dataset needs to be created, it will be produced with only the specified regions.}
-
-\item{prj:}{prject file}
-
-\item{desired_queries:}{desired queries to be loaded}
-}
-\value{
-project file with extra queries
-}
-\description{
-Create a folder to save the datasets and file, in case it does not exist
-}
