+++ NAMESPACE
@@ -21,7 +21,6 @@ export(do_collapse_df)
 export(do_data_sample)
 export(do_mount_tree)
 export(do_unmount_tree)
-export(fill_queries)
 export(filter_data_regions)
 export(filter_variables)
 export(firstup)
