+++ R/main.R
@@ -65,59 +65,6 @@ data_query = function(type, db_path, db_name, prj_name, scenarios, desired_regio
 }
 
 
-#' fill_queries
-#'
-#' Create a folder to save the datasets and file, in case it does not exist
-#' @param db_path: path of the database
-#' @param db_name: name of the database
-#' @param prj_name: name of the project
-#' @param scenarios: name of the scenarios to be considered
-#' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
-#' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
-#' In case the project dataset needs to be created, it will be produced with only the specified regions.
-#' @param prj: prject file
-#' @param desired_queries: desired queries to be loaded
-#' @return project file with extra queries
-#' @export
-fill_queries = function(db_path, db_name, prj_name, scenarios,
-                        desired_regions = 'All', prj, desired_queries) {
-  # add nonCO2 queries manually (they are too big to use the usual method)
-  if (!'nonCO2 emissions by sector (excluding resource production)' %in% rgcam::listQueries(prj) &
-      'nonCO2 emissions by sector (excluding resource production)' %in% desired_queries) {
-    print('nonCO2 emissions by sector (excluding resource production)')
-    dt_sec = data_query('nonCO2 emissions by sector (excluding resource production)', db_path, db_name, prj_name, scenarios, desired_regions)
-    prj_tmp <- rgcam::addQueryTable(project = 'prj_tmp1.dat', qdata = dt_sec, saveProj = FALSE,
-                                    queryname = 'nonCO2 emissions by sector (excluding resource production)', clobber = TRUE)
-    prj <<- rgcam::mergeProjects(prj_name, list(prj,prj_tmp), clobber = TRUE, saveProj = FALSE)
-
-  }
-  if (!'nonCO2 emissions by region' %in% rgcam::listQueries(prj) &
-      'nonCO2 emissions by region' %in% desired_queries) {
-    print('nonCO2 emissions by region')
-    dt_reg = data_query('nonCO2 emissions by region', db_path, db_name, prj_name, scenarios, desired_regions)
-    prj_tmp <- rgcam::addQueryTable(project = 'prj_tmp2.dat', qdata = dt_reg, saveProj = FALSE,
-                                    queryname = 'nonCO2 emissions by region', clobber = TRUE)
-    prj <<- rgcam::mergeProjects(prj_name, list(prj,prj_tmp), clobber = TRUE, saveProj = FALSE)
-  }
-
-  # fix CO2 prices if needed
-  if (!'CO2 prices' %in% rgcam::listQueries(prj) &
-      'CO2 prices' %in% desired_queries) {
-    l = length(rgcam::listScenarios(prj))
-    dt = data.frame(Units = rep(NA,l),
-                    scenario = rgcam::listScenarios(prj),
-                    year = rep(NA,l),
-                    market = rep(NA,l),
-                    value = rep(NA,l))
-    prj_tmp <- rgcam::addQueryTable(project = prj_name, qdata = dt,
-                                    queryname = 'CO2 prices', clobber = TRUE)
-    prj <<- rgcam::mergeProjects(prj_name, list(prj,prj_tmp), clobber = TRUE, saveProj = FALSE)
-  }
-
-  return(prj)
-}
-
-
 #' load_project
 #'
 #' Load specified project into the global environment
@@ -184,7 +131,9 @@ create_project = function(db_path, db_name, prj_name, scenarios,
 
     # read the queries file
     queryFile = paste0('inst/extdata/queries/','queries_gcamreport_gcam7.0_complete.xml')
-    queries_original <- rgcam::parse_batch_query(queryFile)
[…]
