### 2023-12-05 b90ec0c
Minor fixes and filter_variables fix
