+++ R/functions.R
@@ -7,6 +7,19 @@ options(dplyr.summarise.inform = FALSE)
 #########################################################################
 
 
+#' start_with_pattern
+#'
+#' Return the vector elements starting with the specified parameters
+#' @param vector: vector to check
+#' @param pattern: pattern to consider
+#' @return subvector of `vector` whose elements start with `pattern`
+#' @export
+start_with_pattern = function(vector, pattern) {
+  matching_elements = vector[substr(vector, 1, nchar(pattern)) == pattern]
+  return(matching_elements)
+}
+
+
 #' filter_loading_regions
 #'
 #' Filter the desired regions of a GCAM project
@@ -16,6 +29,7 @@ options(dplyr.summarise.inform = FALSE)
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @param variable: dataset variable information
 #' @return filtered dataframe
+#' @export
 filter_loading_regions <- function (data, desired_regions = 'All', variable) {
   if (!(length(desired_regions) == 1 && desired_regions == 'All')) {
     # the variable CO2 prices does not contain "region", but "markets". Now we
@@ -51,13 +65,17 @@ filter_loading_regions <- function (data, desired_regions = 'All', variable) {
 #'
 #' Filter the desired regions of a GCAM project
 #' @param data: dataframe to be filtered
+#' @param variable: variable that requires this data
 #' @return filtered dataframe
 #' @export
-filter_variables = function(data) {
-  if (!(length(desired_variables) == 1 && desired_variables == 'All')) {
-    if ('var' %in% colnames(data)) {
-      data = data %>%
-        dplyr::filter(var %in% desired_variables)
+filter_variables = function(data, variable) {
+
+  if (variable %in% variables[variables$required == TRUE,]$name) {
+    if (!(length(desired_variables) == 1 && desired_variables == 'All')) {
+      if ('var' %in% colnames(data)) {
+        data = data %>%
+          dplyr::filter(var %in% desired_variables)
+      }
     }
   }
 
@@ -297,7 +315,7 @@ get_co2_concentration = function() {
 get_co2 = function() {
   co2_clean <<-
     tibble::as_tibble(rgcam::getQuery(prj, "CO2 emissions by sector (no bio) (excluding resource production)")) %>%
-    dplyr::left_join(filter_variables(co2_sector_map), by = "sector", multiple = "all") %>%
+    dplyr::left_join(filter_variables(co2_sector_map, 'co2_clean'), by = "sector", multiple = "all") %>%
     dplyr::mutate(value = value * unit_conv) %>%
     dplyr::group_by(scenario, region, year, var) %>% #
     dplyr::summarise(value = sum(value, na.rm = T)) %>%
@@ -324,7 +342,7 @@ get_co2_ets = function() {
     tibble::as_tibble(rgcam::getQuery(prj, "nonCO2 emissions by sector (excluding resource production)")) %>%
     dplyr::filter(ghg == 'CO2_ETS') %>%
     # change units to CO2 equivalent and group by sector
-    dplyr::left_join(filter_variables(co2_ets_sector_map), by = "sector", multiple = "all") %>%
+    dplyr::left_join(filter_variables(co2_ets_sector_map, 'co2_ets_bysec'), by = "sector", multiple = "all") %>%
     dplyr::mutate(value = value * unit_conv) %>%
     dplyr::group_by(scenario, region, year, var) %>% #
[…]
