### 2023-12-05 94691be
Minor fixes and filter_variables fix
