+++ NAMESPACE
@@ -22,6 +22,7 @@ export(do_data_sample)
 export(do_mount_tree)
 export(do_unmount_tree)
 export(filter_data_regions)
+export(filter_loading_regions)
 export(filter_variables)
 export(firstup)
 export(gather_map)
@@ -97,5 +98,6 @@ export(load_variable)
 export(reset_first_load)
 export(run)
 export(standard_tech_group)
+export(start_with_pattern)
 export(update_user_choices_plot)
 importFrom(magrittr,"%>%")
