### 2023-12-11 bf8629a
Update and add tests for gcam 7.0
