+++ tests/testthat/testReporter_v7.0.R
@@ -1,0 +1,293 @@
+library(gcamreport); library(testthat); library(magrittr); library(rprojroot); library(rpackageutils)
+
+test_that("Test1_v7. download db, create project, and run", {
+  # load a reference GCAM db form a Zenodo repository
+  db_path = paste0(rprojroot::find_root(rprojroot::is_testthat),'/testInputs/v_7.0')
+  rpackageutils::download_unpack_zip(data_directory = db_path,
+                                     url = "https://zenodo.org/record/10258919/files/database_basexdb_ref.zip?download=1")
+  testthat::expect_equal(1, 1)
+
+  # create the prj
+  db_name = "database_basexdb_ref"
+  prj_name = "gcamv7.0_test.dat"
+  scenarios = 'Reference'
+
+  create_project(db_path, db_name, prj_name, scenarios)
+  prj_tmp = prj
+  testResult = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test7.dat')))
+  prj = prj_tmp
+  print('................test queries')
+  testthat::expect_equal(prj$Reference$`nonCO2 emissions by region`, testResult$Reference$`nonCO2 emissions by region`)
+  testthat::expect_equal(prj$Reference$`nonCO2 emissions by sector`, testResult$Reference$`nonCO2 emissions by sector (excluding resource production)`)
+  testthat::expect_equal(prj$Reference$`CO2 prices`, testResult$Reference$`CO2 prices`)
+
+  # check nonCO2 emissions query
+  dt_sec = data_query('nonCO2 emissions by sector (excluding resource production)', db_path, db_name, prj_name, scenarios)
+  testResult = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/result_test1.RData')))
+  testthat::expect_equal(dt_sec, testResult)
+
+  # clean environment
+  rm(list = ls())
+})
+
+test_that("Test2_v7. load project", {
+  testResult = as.numeric(length(load_project(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testInputs/v_7.0/test7.dat'))))
+  testthat::expect(!is.null(testResult), 'Null project. Check if the path exists or the "load_project" function works correctly.')
+
+  # clean environment
+  rm(list = ls())
+})
+
+test_that("Test3_v7. run - dataset created", {
+  run(project_path = paste0(rprojroot::find_root(rprojroot::is_testthat),'/testInputs/v_7.0/test7.dat'), launch_ui = FALSE)
+  testthat::expect(!is.null(final_data) & dplyr::n_distinct(final_data) > 0, 'Empty dataset. Check if the project path exists or the "run" function works correctly.')
+
+  # clean environment
+  rm(list = ls())
+})
+
+test_that("Test4_v7. run - dataset saved with output_file specified", {
+  run(project_path = paste0(rprojroot::find_root(rprojroot::is_testthat),'/testInputs/v_7.0/test7.dat'), launch_ui = FALSE,
+      output_file = paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test7_output'))
+
+  testResult1 = read.csv(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test7_output.csv'))
+  testthat::expect(dplyr::n_distinct(testResult1) > 0, 'Dataset not saved. Check if the project path exists or the "run" function works correctly.')
+  testResult2 = read.csv(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/result_test4.1.csv'))
+  testthat::expect_equal(testResult1, testResult2)
+
[…]
