+++ .github/workflows/test_coverage.yml
@@ -40,16 +40,9 @@ jobs:
           install.packages(c("remotes","rcmdcheck","covr"), repos = "https://cloud.r-project.org")
           remotes::install_github("JGCRI/rpackageutils")
           remotes::install_deps(dependencies = TRUE)
+          install.packages(c("DT", "htmltools"))
         shell: Rscript {0}
 
       - name: Test coverage
-        run: |
-          cov <- covr::package_coverage()
-          covr:::report(cov, file = "coverage.xml")
+        run: covr::codecov(token = "${{secrets.CODECOV_TOKEN}}")
         shell: Rscript {0}
-
-      - uses: codecov/codecov-action@v3
-        with:
-          token: ${{secrets.CODECOV_TOKEN}}
-          files: ./coverage.xml
-          verbose: true
