+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
   Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
   Version_Guide: Version_Guide.html
-last_built: 2023-12-11T11:43Z
+last_built: 2023-12-11T14:08Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
