### 2023-12-11 bc3LC/gcamreport#5 @bpbond
Code suggestions
https://github.com/openjournals/joss-reviews/issues/5975

- [x] Many parts of the code use `paste0` to construct file paths...R's [file.path](https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/file.path) function is a more robust way to do this
- [x] Most R style guides, including Google and [Tidyverse](https://style.tidyverse.org/syntax.html#assignment-1), recommend using `<-` for assignment, not `=`
- [x] Consider running the package through the `styler` package, with a style of your choice, to ensure clear and consistent code style
- [x] `rm(list = ls())` is not needed at the end of test blocks; `testthat` takes care of cleanup
- [x] `library` calls should never appear in package code
- [x] Since you're using Roxygen, use `importFrom` in function headers to make functions available (and eliminate all the `dplyr::`-type namespace qualifiers, too, improving readability)
- [x] On my system, checking the package currently generates many NOTEs and WARNINGs: undocumented function arguments,  non-ASCII strings, missing documentation, unrecognized documentation tags, undefined/no visible binding variables
- [x] Package tests generate lots of printed outputs, which makes it hard to see test results
