### 2023-12-11 #5975 @bpbond
My apologies to @martinfleis and the authors for the delay in my review.

This is an interesting contribution with the potential to be a broadly useful resource and source of documentation. I have read the manuscript, and installed the software and examined its code; see specific suggestions about both referenced above. Broadly:
* The manuscript is generally well written but needs many minor clarifications, and would be improved by a specific example and more-readable figure
* I highly recommend including a sample `awesomeProject.dat` dataset so that users (and reviewers) can easily test out the software _without_ running GCAM
* Running R CMD CHECK on the code currently generates many notes, warnings, and errors. Maybe not everything can be solved, but at least the authors should follow basic best practices for packages with respect to namespaces and dependencies, completeness of documentation, etc.

I hope this is useful! I appreciate that the authors have done a lot of work, and am happy to clarify/answer questions if useful.

**Conflict of interest**: please note that I work in the same [institute](https://www.pnnl.gov/projects/jgcri) as two of the co-authors. I don't work with them directly, and don't feel conflicted in giving this review, but it's important to note transparently.
