+++ man/generate_report.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/main.R
-\name{run}
-\alias{run}
-\title{run}
+\name{generate_report}
+\alias{generate_report}
+\title{generate_report}
 \usage{
-run(
+generate_report(
   project_path = NULL,
   db_path = NULL,
   db_name = NULL,
@@ -63,7 +63,7 @@ RData, CSV, and XLSX datafile with the desired variables & launches user interfa
 \description{
 Main function. Creates/loads a GCAM project, standardizes the data and saves it
 in several forms (RData, CSV, and XLSX), runs vetting verifications, and launches
-the User Interface. Specify regions and/or variables to be reported and point
+the User Interface. You can specify the regions and/or variables to be reported and point
 either the `project_path`, or the `db_path`, the `db_name`, the `prj_name`, and `scenarios`.
-The resulting RData output can be used to call manually `launch_gcamreport_ui`.
+The resulting RData output can be used to manually call `launch_gcamreport_ui`.
 }
