+++ man/conv_ghg_co2e.Rd
@@ -6,6 +6,9 @@
 \usage{
 conv_ghg_co2e(data)
 }
+\arguments{
+\item{data:}{dataset}
+}
 \description{
 Covert GHG to CO2e
 }
