### 2023-12-13 9ad36c7
Fix test6: test_coverage
