### 2023-12-13 15b1823
Fix test7: test_coverage
