+++ pkgdown.yml
@@ -3,10 +3,15 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   Dataset_Generation_Tutorial: Dataset_Generation_Tutorial.html
+  Dataset_Generation_Tutorial_v6-0: Dataset_Generation_Tutorial_v6-0.html
   Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
+  Interactive_UI_Tutorial_v6-0: Interactive_UI_Tutorial_v6-0.html
+  README_v6-0: README_v6-0.html
   Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
+  Step_By_Step_Full_Example_v6-0: Step_By_Step_Full_Example_v6-0.html
   Version_Guide: Version_Guide.html
-last_built: 2023-12-13T10:23Z
+  gcamreport_v6-0: gcamreport_v6-0.html
+last_built: 2023-12-13T12:53Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
