### 2023-12-13 aa1dade
Update paper with Ben Bond-Lamberty inputs
