+++ paper/paper.md
@@ -39,34 +39,39 @@ bibliography: paper.bib
 
 # Summary
 
-There is an urgent need to perform multi-model studies to deal with models’ heterogeneity and uncertainty with the aim of building a more reliable and transparent framework to inform policy makers in the design and implementation of climate policies [@guivarch_using_2022]. In response to this challenge, multiple institutes and organizations have adopted the standardized data template developed by the Integrated Assessment Modeling Consortium (IAMC). This template is maintained by the International Institute for Applied Systems Analysis (IIASA) and aims to standardize and facilitate model intercomparison exercises. For the latest Assessment Report (AR6), the Intergovernmental Panel on Climate Change (IPCC) required all contributors to homogenize their data to enable comparisons and ensure full transparency [@krey_annex_2014]. This practice has set the foundation for a new open management of the outputs in the area of global scenario analysis.  
+There is an urgent need to perform multi-model studies to characterize uncertainty arising from models' heterogeneity with the aim of building a more reliable and transparent framework to inform policy makers in the design and implementation of climate policies [@guivarch_using_2022]. In response to this challenge, multiple institutes and organizations have adopted the standardized data template developed by the Integrated Assessment Modeling Consortium (IAMC). This template is maintained by the International Institute for Applied Systems Analysis (IIASA) and aims to standardize and facilitate model intercomparison exercises. For the latest Assessment Report (AR6), the Intergovernmental Panel on Climate Change (IPCC) required all contributors to homogenize their data to enable comparisons and ensure full transparency [@krey_annex_2014]. This practice has set the foundation for a new open management of the outputs in the area of global scenario analysis.  
 
-In the case of the Global Change Analysis Model (GCAM) [@calvin2019], a well-reputed model that has been extensively used by different international and national scenario analysis, the harmonization code has never been documented nor standardized, making it difficult to reproduce the outputs and hindering the transparency of the results. To overcome these limitations, we have developed `gcamreport`, an R package that systematizes the transformations of GCAM outputs, generates figures to facilitate the analysis of the results, and allows user interaction with the produced outputs. Furthermore, the tool can be used embedded in a Docker image, which allows users to use the package in a virtual environment without having to install any specific software or library. Finally, each `gcamreport` release is linked to either a version of GCAM or a study in which GCAM was used, ensuring reproducibility, interoperability, accessibility, and findability, which is in line with the well-known open science principles FAIR and TRUST [@wilkinson2016; @lin2020].
+In the case of the Global Change Analysis Model (GCAM) [@calvin2019], a well-regarded model that has been extensively used for different international and national scenario analysis, the harmonization code has never been documented nor standardized, making it difficult to reproduce outputs and hindering the transparency of results. To overcome these limitations, we have developed `gcamreport`, an R package that systematizes the transformations of GCAM outputs, generates figures to facilitate the analysis of the results, and allows user interaction with the produced outputs. Furthermore, the tool can be used embedded in a Docker image, which allows users to use the package in a virtual environment without having to install any specific software or library. Finally, each `gcamreport` release is linked to either a version of GCAM or a study in which GCAM was used, ensuring reproducibility, interoperability, accessibility, and findability, which is in line with the well-known open science principles FAIR and TRUST [@wilkinson2016; @lin2020].
 
 
-![Figure 1: Structure of the `gcamreport` package.\label{fig:scheme}](scheme_fig3.png){width="75%"}
+![**Figure 1:** Structure of the `gcamreport` package.\label{fig:scheme}](scheme_fig4_crop.png){width="75%"}
 
[…]
