+++ pkgdown.yml
@@ -11,7 +11,7 @@ articles:
   Step_By_Step_Full_Example_v6-0: Step_By_Step_Full_Example_v6-0.html
   Version_Guide: Version_Guide.html
   gcamreport_v6-0: gcamreport_v6-0.html
-last_built: 2023-12-13T12:53Z
+last_built: 2023-12-13T14:53Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
