+++ tests/testthat/testUI.R
@@ -1,12 +1,15 @@
-library(gcamreport); library(testthat); library(magrittr); library(rprojroot)
+library(gcamreport); library(testthat); library(magrittr)
 
 test_that("Test1. test tree functions", {
-  generate_report(project_path = paste0(rprojroot::find_root(rprojroot::is_testthat),'/testInputs/v_7.0/test7.dat'), launch_ui = FALSE)
+
+  generate_report(project_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/test7.dat'), launch_ui = FALSE)
   # load data
-  data = report
+  data <- report
   # define the dataset for launching the ui
-  sdata <<- data %>%
-    tidyr::separate(Variable, into = c('col1','col2','col3','col4','col5','col6','col7'), sep = "([\\|])", extra = 'merge', remove = FALSE)
+  sdata <<- suppressWarnings(
+    data %>%
+      tidyr::separate(Variable, into = c('col1','col2','col3','col4','col5','col6','col7'), sep = "([\\|])", extra = 'merge', remove = FALSE)
+    )
   # create vector of available years for launching the ui
   available_years <<- as.numeric(names(sdata)[13:length(names(sdata))])
   # develop a nested list of the variables and regions for launching the ui
@@ -18,47 +21,48 @@ test_that("Test1. test tree functions", {
 
 
   # do_mount_tree with regions
-  testResult1 = do_mount_tree(reg_cont,names(reg_cont),selec=TRUE)
-  testExpect1 = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test_ui_1.RData')))
+  testResult1 <- do_mount_tree(reg_cont,names(reg_cont),selec=TRUE)
+  testExpect1 <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat),'testOutputs/v_7.0/test_ui_1.RData')))
   testthat::expect_equal(testResult1, testExpect1)
 
   # do_mount_tree with variables
-  testResult2 = do_mount_tree(cols,names(cols),selec=TRUE)
-  testExpect2 = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test_ui_2.RData')))
+  testResult2 <- do_mount_tree(cols,names(cols),selec=TRUE)
+  testExpect2 <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat),'testOutputs/v_7.0/test_ui_2.RData')))
   testthat::expect_equal(testResult2, testExpect2)
 
   # do_collapse_df
-  testResult3 = do_collapse_df(cols)
-  testExpect3 = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test_ui_3.RData')))
+  testResult3 <- do_collapse_df(cols)
+  testExpect3 <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat),'testOutputs/v_7.0/test_ui_3.RData')))
   testthat::expect_equal(testResult3, testExpect3)
 
   # change_style
-  test4_pre = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test_ui_4_pre.RData')))
-  testExpect4 = get(load(paste0(rprojroot::find_root(rprojroot::is_testthat),'/testOutputs/v_7.0/test_ui_4_post.RData')))
-  testResult4 = change_style(test4_pre, 'regions')
[…]
