+++ R/app_functions.R
@@ -6,13 +6,13 @@
 #' compute_height
 #'
 #' Return perfect heigh of rendered plots with ungrouped regions
-#' @param reg_all: regions' vector
+#' @param reg_all regions' vector
 #' @keywords internal
-#' @importFrom magrittr %>%
+#' @importFrom dplyr if_else
 #' @export
 compute_height <- function(reg_all) {
-  reg_clean = setdiff(reg_all, c('MAF','LAM','OECD90','REF','ASIA'))
-  h = dplyr::if_else(length(reg_clean) < 3, 325, 160*ceiling(sqrt(length(reg_clean))))
+  reg_clean <- setdiff(reg_all, c('MAF','LAM','OECD90','REF','ASIA'))
+  h <- if_else(length(reg_clean) < 3, 325, 160*ceiling(sqrt(length(reg_clean))))
 
   return(h)
 }
@@ -21,9 +21,8 @@ compute_height <- function(reg_all) {
 #' is_leaf
 #'
 #' Return TRUE if it is a leaf node; FALSE otherwise.
-#' @param tree: tree node to be checked
+#' @param tree tree node to be checked
 #' @keywords internal
-#' @importFrom magrittr %>%
 #' @export
 is_leaf <- function(tree) {
   if (length(tree) > 1 || length(tree[[1]]) > 1 || length(tree[[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]]) > 1
@@ -38,9 +37,9 @@ is_leaf <- function(tree) {
 #' change_style
 #'
 #' Change the style of the tree nodes: 'dis' if needs to be disabled, 'basic' otherwise.
-#' @param tree: base tree to change the style on
-#' @param type: either regions or variables
-#' @param tmp_vars: variables which require 'dis' style
+#' @param tree base tree to change the style on
+#' @param type either regions or variables
+#' @param tmp_vars variables which require 'dis' style
 #' @keywords internal
 #' @importFrom magrittr %>%
 #' @export
@@ -59,9 +58,9 @@ change_style <- function(tree, type, tmp_vars = NULL) {
     # change the style for the current node. If it's in the list of variables to be disabled, set
     # it to 'dis'. To 'basic' otherwise
     if (type == 'variables' && length(tmp_vars) > 0 && attrs$my_id %in% tmp_vars) {
-      attrs$sttype = 'dis'
+      attrs$sttype <- 'dis'
     } else {
-      attrs$sttype = 'basic'
+      attrs$sttype <- 'basic'
     }
     attributes(tree[[i]]) <- attrs
   }
@@ -74,24 +73,25 @@ change_style <- function(tree, type, tmp_vars = NULL) {
 #'
 #' Check user's choices to do the plot: at least one scenario, variable, year, and region
 #' must be chosen. In case of 'grouped' plot, all variables must be from the same category.
-#' @param scen: user's selected scenarios
-#' @param years: user's selected years
-#' @param reg: user's selected regions
-#' @param vars: user's selected variables
-#' @param grouped: if TRUE, aim to display grouped plot; ungrouped plot otherwise
+#' @param scen user's selected scenarios
+#' @param years user's selected years
+#' @param reg user's selected regions
+#' @param vars user's selected variables
+#' @param grouped if TRUE, aim to display grouped plot; ungrouped plot otherwise
 #' @keywords internal
 #' @importFrom magrittr %>%
+#' @importFrom stringr str_extract
 #' @export
 check_user_choices_plot <- function(vars, scen, years, reg, grouped) {
 
   # errors' vector
-  error_message = c()
+  error_message <- c()
 
   # study the category of the selected variables
   if (length(vars) > 0) {
[…]
