+++ man/get_co2_price_share.Rd
@@ -14,5 +14,4 @@ Get co2 price between CO2 and CO2_ETS. If only one CO2 type present, share = 1;
 otherwise, each type has the share corresponding to the last historical year
 }
 \keyword{co2}
-\keyword{rgcam::getQuery}
 \keyword{tmp}
