+++ Dockerfile
@@ -26,6 +26,7 @@ RUN R -e "install.packages('readr')"
 RUN R -e "install.packages('data.table')"
 RUN R -e "install.packages('magrittr')"
 RUN R -e "install.packages('dplyr')"
+RUN R -e "install.packages('tibble')"
 RUN R -e "install.packages('tidyr')"
 RUN R -e "install.packages('ggplot2')"
 RUN R -e "install.packages('stringr')"
