### 2023-12-19 a8c803c
Review: fix global variables names
