### 2023-12-19 3f383e4
Review: fix global variables names
