+++ R/main.R
@@ -90,7 +90,7 @@ load_project <- function(project_path, desired_regions = 'All') {
       }
     }
   }
-  Scenarios <<- listScenarios(prj)
+  scenarios.global <<- listScenarios(prj)
 
   prj <<- prj
 }
@@ -238,7 +238,7 @@ create_project <- function(db_path, db_name, prj_name, scenarios,
     saveProject(prj, file = file.path(db_path, paste(db_name, prj_name, sep = '_')))
 
 
-    Scenarios <<- listScenarios(prj)
+    scenarios.global <<- listScenarios(prj)
 
     prj <<- prj
   }
@@ -262,7 +262,7 @@ load_variable <- function(var){
   # if the variable has dependencies, load them
   if (!is.na(var$dependencies)) {
     for (d in var$dependencies[[1]]) {
-      load_variable(variables[which(variables$name == d),])
+      load_variable(variables.global[which(variables.global$name == d),])
     }
   }
 
@@ -273,7 +273,7 @@ load_variable <- function(var){
   get(var$fun)()
 
   # keep record of the loaded variables
-  loaded_internal_variables <<- c(loaded_internal_variables, var$name)
+  loaded_internal_variables.global <<- c(loaded_internal_variables.global, var$name)
 }
 
 
@@ -401,6 +401,7 @@ available_variables <- function(print = TRUE) {
 #' @return RData, CSV, and XLSX datafile with the desired variables & launches user interface.
 #' @import dplyr
 #' @importFrom tidyr replace_na
+#' @importFrom utils write.csv
 #' @importFrom writexl write_xlsx
 #' @importFrom stringr str_sub str_locate
 #' @export
@@ -503,11 +504,11 @@ generate_report <- function(project_path = NULL, db_path = NULL, db_name = NULL,
     stop('ERROR: Specify either a project or a database to extract the data from.')
   }
 
-  # make final_db_year as a global variable
-  final_db_year <<- final_year
+  # make final_year as a global variable
+  final_year.global <<- final_year
 
   # final reporting columns
-  reporting_columns_fin <<- append(c("Model", "Scenario", "Region", "Variable", "Unit"), as.character(seq(2005, final_db_year, by = 5)))
+  reporting_columns.global <<- append(c("Model", "Scenario", "Region", "Variable", "Unit"), as.character(seq(2005, final_year.global, by = 5)))
 
   # desired variables to have in the report
   variables_base <- data.frame('name' = unique(template$Internal_variable)[!is.na(unique(template$Internal_variable)) & unique(template$Internal_variable) != ""],
@@ -516,9 +517,9 @@ generate_report <- function(project_path = NULL, db_path = NULL, db_name = NULL,
 
   # consider only the desired variables
   if (length(desired_variables) == 1 && desired_variables == 'All') {
-    variables <<- variables_base
+    variables.global <<- variables_base
   } else {
-    variables <<- variables_base %>%
+    variables.global <<- variables_base %>%
       mutate(required = if_else(
         !name %in% unique(template %>%
                             filter(Variable %in% desired_variables) %>%
@@ -529,16 +530,16 @@ generate_report <- function(project_path = NULL, db_path = NULL, db_name = NULL,
   print('Loading data, performing checks, and saving output...')
 
   # consider the dependencies and checking functions
-  variables <<- merge(variables,var_fun_map, by = 'name', all = TRUE) %>%
+  variables.global <<- merge(variables.global,var_fun_map, by = 'name', all = TRUE) %>%
     replace_na(list(required = FALSE))
 
   # for all desired variables, load the corresponding data
[…]
