### 2023-12-19 5ce6127
Review: fix global data items warnings and clean unnecessary items
