+++ R/data.R
@@ -418,125 +418,15 @@
 #' }
 "co2_market_frag_map"
 
-#' CUMULATIVE_EMISSIONS_BUDGET_GOALS
+#' convert
 #'
-#' @description cumulative emissions budget goals list
+#' @description units conversion list
 #' @examples
 #' \dontrun{
 #'  library(gcamreport);
-#'  gcamreport::CUMULATIVE_EMISSIONS_BUDGET_GOALS
+#'  gcamreport::convert
 #' }
-"CUMULATIVE_EMISSIONS_BUDGET_GOALS"
-
-#' conv_thousand_million
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_thousand_million
-#' }
-"conv_thousand_million"
-
-#' conv_million_billion
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_million_billion
-#' }
-"conv_million_billion"
-
-#' conv_90USD_10USD
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_90USD_10USD
-#' }
-"conv_90USD_10USD"
-
-#' conv_75USD_10USD
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_75USD_10USD
-#' }
-"conv_75USD_10USD"
-
-#' conv_15USD_10USD
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_15USD_10USD
-#' }
-"conv_15USD_10USD"
-
-#' conv_19USD_75USD
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_19USD_75USD
-#' }
-"conv_19USD_75USD"
-
-#' conv_C_CO2
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::conv_C_CO2
-#' }
-"conv_C_CO2"
-
-#' hr_per_yr
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::hr_per_yr
-#' }
-"hr_per_yr"
-
-#' EJ_to_GWh
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::EJ_to_GWh
-#' }
-"EJ_to_GWh"
-
-#' bcm_to_EJ
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::bcm_to_EJ
-#' }
-"bcm_to_EJ"
-
-#' GJ_to_EJ
-#'
-#' @description units conversion
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::GJ_to_EJ
-#' }
-"GJ_to_EJ"
+"convert"
 
 #' F_GASES
 #'
@@ -607,134 +497,4 @@
 #'  gcamreport::long_columns
 #' }
 "long_columns"
-
-#' reporting_columns
-#'
-#' @description reporting columns
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
-#'  gcamreport::reporting_columns
-#' }
-"reporting_columns"
-
-#' state_names
-#'
-#' @description state names
-#' @examples
-#' \dontrun{
-#'  library(gcamreport);
[…]
