+++ man/F_GASES.Rd
@@ -5,7 +5,7 @@
 \alias{F_GASES}
 \title{F_GASES}
 \format{
-An object of class \code{character} of length 13.
+An object of class \code{character} of length 16.
 }
 \usage{
 F_GASES
