+++ man/GJ_to_EJ.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{GJ_to_EJ}
-\alias{GJ_to_EJ}
-\title{GJ_to_EJ}
-\format{
-An object of class \code{numeric} of length 1.
-}
-\usage{
-GJ_to_EJ
-}
-\description{
-units conversion
-\dontrun{
- library(gcamreport);
- gcamreport::GJ_to_EJ
-}
-}
-\keyword{datasets}
