+++ man/convert.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{convert}
+\alias{convert}
+\title{convert}
+\format{
+An object of class \code{list} of length 12.
+}
+\usage{
+convert
+}
+\description{
+<<<<<<< HEAD:man/GJ_to_EJ.Rd
+units conversion
+=======
+units conversion list
+}
+\examples{
+>>>>>>> f4b3e58 (Review: fix global data items warnings and clean unnecessary items):man/convert.Rd
+\dontrun{
+ library(gcamreport);
+ gcamreport::convert
+}
+}
+\keyword{datasets}
