+++ NAMESPACE
@@ -21,6 +21,7 @@ export(do_unmount_tree)
 export(filter_data_regions)
 export(filter_loading_regions)
 export(filter_variables)
+export(gather_map)
 export(generate_report)
 export(get_CCS_invest)
 export(get_ag_demand)

[…]
