### 2023-12-19 f4b3e58
Review: fix global data items warnings and clean unnecessary items
