+++ inst/extdata/saveDataFiles.R
@@ -185,59 +185,36 @@ use_data(co2_market_frag_map, overwrite=T)
 
 # List of Constants
 
-CUMULATIVE_EMISSIONS_BUDGET_GOALS <- c(300, 400, 500, 600, 700, 800, 900, 950, 1000, 1200, 1400, 1600, 1800, 2000, 2500, 3000)
-use_data(CUMULATIVE_EMISSIONS_BUDGET_GOALS, overwrite=T)
-
-# Basic format conv_[from]_[to]
-conv_thousand_million <- 1/1000
-use_data(conv_thousand_million, overwrite=T)
-
-conv_million_billion <- 1/1000
-use_data(conv_million_billion, overwrite=T)
-
-# NOTE: These values are only used for queries that don't have an associated mapping file
-# for queries such as primary_fuel_prices this conversion is specified in the mapping file
-# These values are taken from GDP inflator in the GCAM R package
-conv_90USD_10USD <- 1.515897
-use_data(conv_90USD_10USD, overwrite=T)
-
-conv_75USD_10USD <- 3.227608
-use_data(conv_75USD_10USD, overwrite=T)
-
-conv_15USD_10USD <- .91863
-use_data(conv_15USD_10USD, overwrite=T)
-
-conv_19USD_75USD <- .2658798
-use_data(conv_19USD_75USD, overwrite=T)
-
-conv_C_CO2 <- 44/12
-use_data(conv_C_CO2, overwrite=T)
-
-
-# Elec related conversions
-hr_per_yr <- 8760
-use_data(hr_per_yr, overwrite=T)
-
-EJ_to_GWh <- 0.0000036
-use_data(EJ_to_GWh, overwrite=T)
-
-bcm_to_EJ <- 0.03600
-use_data(bcm_to_EJ, overwrite=T)
-
-GJ_to_EJ <- 1.0E9
-use_data(GJ_to_EJ, overwrite=T)
-
+convert <- list(
+  # Basic format conv_[from]_[to]
+  conv_thousand_million = 1/1000,
+  conv_million_billion = 1/1000,
+  # NOTE: These values are only used for queries that don't have an associated mapping file
+  # for queries such as primary_fuel_prices this conversion is specified in the mapping file
+  # These values are taken from GDP inflator in the GCAM R package
+  conv_90USD_10USD = 1.515897,
+  conv_75USD_10USD = 3.227608,
+  conv_15USD_10USD = 0.91863,
+  conv_19USD_75USD = 0.2658798,
+  conv_C_CO2 = 44/12,
+  # Elec related conversions
+  hr_per_yr = 8760,
+  EJ_to_GWh = 0.0000036,
+  bcm_to_EJ = 0.03600,
+  GJ_to_EJ = 1.0E9,
+  #ghg * CO2_equivalent gives CO2 units
+  CO2_equivalent = 3.666667
+)
+use_data(convert, overwrite=T)
 
 # GHG emission conversion
 F_GASES <- c("C2F6", "CF4", "HFC125", "HFC134a", "HFC245fa", "SF6", "HFC143a",
-             "HFC152a", "HFC227ea", "HFC23", "HFC236fa", "HFC32", "HFC365mfc")
+             "HFC152a", "HFC227ea", "HFC23", "HFC236fa", "HFC32", "HFC365mfc",
+             "HFC43", "HFC245fa", "HFC43-10")
 use_data(F_GASES, overwrite=T)
 
 GHG_gases <- c("CH4", "N2O", F_GASES, "CO2", "CO2LUC")
 use_data(GHG_gases, overwrite=T)
-
-CO2_equivalent <- 3.666667 #ghg * CO2_equivalent gives CO2 units
-use_data(CO2_equivalent, overwrite=T)
 
 
[…]
