+++ R/functions.R
@@ -85,6 +85,21 @@ filter_variables <- function(data, variable) {
   return(invisible(data))
 }
 
+#' gather_map
+#'
+#' Formats all maps into a long table
+#' @keywords internal
+#' @importFrom tidyr gather
+#' @importFrom dplyr select filter
+#' @export
+gather_map <- function(df){
+  untouched_cols <- names(df) %>% .[!grepl("var", names(df))]
+  df %>%
+    gather(identifier, var, -untouched_cols) %>%
+    select(-identifier) %>%
+    filter(!is.na(var), var != "") %>%
+    return()
+}
 
 #' conv_ghg_co2e
 #'
@@ -95,18 +110,12 @@ filter_variables <- function(data, variable) {
 #' @importFrom magrittr %>%
 #' @export
 conv_ghg_co2e <- function (data) {
-
-  # Ancillary function
   # GHG emission conversion
-  F_GASES <- c("C2F6", "CF4", "HFC125", "HFC134a", "HFC245fa", "SF6", "HFC143a", "HFC152a", "HFC227ea", "HFC23", "HFC236fa", "HFC32", "HFC365mfc", "HFC43",
-               "HFC245fa", "HFC43-10")
-  GHG_gases <- c("CH4", "N2O", F_GASES, "CO2", "CO2LUC")
-
   res <- suppressWarnings(
     data %>%
       separate(ghg, into = c("variable", "sector"), sep = "_", fill = "right") %>%
-      filter(variable %in% GHG_gases) %>%
-      left_join(GWP_adjuster, by = c("variable" = "GHG_gases")) %>%
+      filter(variable %in% gcamreport::GHG_gases) %>%
+      left_join(gcamreport::GWP_adjuster, by = c("variable" = "GHG_gases")) %>%
       mutate(value = value * GWP, Units = "CO2e") %>%
       select(-GWP)
   )
@@ -125,12 +134,8 @@ conv_ghg_co2e <- function (data) {
 #' @importFrom magrittr %>%
 #' @export
 conv_EJ_GW <- function (data, cf, EJ){
-  # Elec related conversions
-  hr_per_yr <- 8760
-  EJ_to_GWh <- 0.0000036
-
   data %>%
-    mutate(gw = EJ / (cf * hr_per_yr * EJ_to_GWh))
+    mutate(gw = EJ / (cf * gcamreport::convert$hr_per_yr * gcamreport::convert$EJ_to_GWh))
 }
 
 #' approx_fun
@@ -177,9 +182,9 @@ approx_fun <- function(year, value, rule = 1) {
 get_population <- function() {
   population_clean <<-
     getQuery(prj, "population by region") %>%
-    mutate(value = value * conv_thousand_million,
+    mutate(value = value * gcamreport::convert$conv_thousand_million,
            var = "Population") %>%
-    select(all_of(long_columns))
+    select(all_of(gcamreport::long_columns))
 }
 
 
@@ -196,9 +201,9 @@ get_gdp_ppp <- function() {
   GDP_PPP_clean <<-
     getQuery(prj, "GDP per capita PPP by region") %>%
     left_join(population_clean %>% rename(pop_mill = value), by = c("scenario", "region", "year")) %>%
-    mutate(value = value * pop_mill * conv_90USD_10USD,
+    mutate(value = value * pop_mill * gcamreport::convert$conv_90USD_10USD,
            var = "GDP|PPP") %>%
-    select(all_of(long_columns))
+    select(all_of(gcamreport::long_columns))
 }
 
 
@@ -214,9 +219,9 @@ get_gdp_ppp <- function() {
 get_gdp_mer <- function() {
   GDP_MER_clean <<-
     getQuery(prj, "GDP MER by region") %>%
[…]
