+++ man/tech.list.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{tech.list}
-\alias{tech.list}
-\title{tech.list}
-\format{
-An object of class \code{character} of length 20.
-}
-\usage{
-tech.list
-}
-\description{
-techonolgies list
-}
-\examples{
-\dontrun{
- library(gcamreport);
- gcamreport::tech.list
-}
-}
-\keyword{datasets}

[…]
