+++ man/get_ag_prices_wld_tmp.Rd
@@ -13,5 +13,9 @@ ag_prices_wld global variable
 Get ag price index.
 }
 \keyword{ag}
+<<<<<<< HEAD
 \keyword{rgcam::getQuery}
+=======
+\keyword{internal}
+>>>>>>> dc7f961 (Review: fix warning message undefined data)
 \keyword{tmp}
