+++ man/get_co2_price_global_tmp.Rd
@@ -13,5 +13,9 @@ co2_price_global & regions global variable
 Get global co2 price.
 }
 \keyword{co2}
+<<<<<<< HEAD
 \keyword{rgcam::getQuery}
+=======
+\keyword{internal}
+>>>>>>> dc7f961 (Review: fix warning message undefined data)
 \keyword{tmp}
