### 2023-12-19 dc7f961
Review: fix warning message undefined data
