+++ R/functions.R
@@ -12,6 +12,7 @@ options(dplyr.summarise.inform = FALSE)
 #' @param pattern pattern to consider
 #' @return subvector of `vector` whose elements start with `pattern`
 #' @importFrom magrittr %>%
+#' @keywords internal utils
 #' @export
 start_with_pattern <- function(vector, pattern) {
   matching_elements <- vector[substr(vector, 1, nchar(pattern)) == pattern]
@@ -30,8 +31,11 @@ start_with_pattern <- function(vector, pattern) {
 #' @return filtered dataframe
 #' @importFrom dplyr mutate filter select
 #' @importFrom magrittr %>%
+#' @keywords internal utils
 #' @export
 filter_loading_regions <- function (data, desired_regions = 'All', variable) {
+  market <- region <- NULL
+
   if (!(length(desired_regions) == 1 && desired_regions == 'All')) {
     # the variable CO2 prices does not contain "region", but "markets". Now we
     # filter for all market items that do not contain the desired regions
@@ -70,8 +74,10 @@ filter_loading_regions <- function (data, desired_regions = 'All', variable) {
 #' @return filtered dataframe
 #' @importFrom dplyr filter
 #' @importFrom magrittr %>%
+#' @keywords internal utils
 #' @export
 filter_variables <- function(data, variable) {
+  var <- NULL
 
   if (variable %in% variables.global[variables.global$required == TRUE,]$name) {
     if (!(length(desired_variables) == 1 && desired_variables == 'All')) {
@@ -85,6 +91,7 @@ filter_variables <- function(data, variable) {
   return(invisible(data))
 }
 
+
 #' gather_map
 #'
 #' Formats all maps into a long table
@@ -93,6 +100,8 @@ filter_variables <- function(data, variable) {
 #' @importFrom dplyr select filter
 #' @export
 gather_map <- function(df){
+  . <- identifier <- var <- NULL
+
   untouched_cols <- names(df) %>% .[!grepl("var", names(df))]
   df %>%
     gather(identifier, var, -untouched_cols) %>%
@@ -101,6 +110,7 @@ gather_map <- function(df){
     return()
 }
 
+
 #' conv_ghg_co2e
 #'
 #' Covert GHG to CO2e
@@ -108,8 +118,11 @@ gather_map <- function(df){
 #' @importFrom dplyr filter left_join mutate select
 #' @importFrom tidyr separate
 #' @importFrom magrittr %>%
+#' @keywords internal conversion
 #' @export
 conv_ghg_co2e <- function (data) {
+  ghg <- variable <- value <- GWP <- NULL
+
   # GHG emission conversion
   res <- suppressWarnings(
     data %>%
@@ -132,6 +145,7 @@ conv_ghg_co2e <- function (data) {
 #' @param EJ EJ amount
 #' @importFrom dplyr mutate
 #' @importFrom magrittr %>%
+#' @keywords internal conversion
 #' @export
 conv_EJ_GW <- function (data, cf, EJ){
   data %>%
@@ -146,6 +160,7 @@ conv_EJ_GW <- function (data, cf, EJ){
 #' @param rule number of points to extrapolate
 #' @importFrom stats approx
 #' @importFrom magrittr %>%
+#' @keywords internal utils
 #' @export
 approx_fun <- function(year, value, rule = 1) {
   if (rule == 1 | rule == 2) {
@@ -173,13 +188,15 @@ approx_fun <- function(year, value, rule = 1) {
 #' get_population
 #'
 #' Get population query and change units to [million].
-#' @keywords population
+#' @keywords internal population
[…]
