+++ man/get_fe_transportation_tmp.Rd
@@ -13,4 +13,5 @@ fe_transportation global variable
 Get mode-specific transport final energy to break out rail, ship, and domestic air.
 }
 \keyword{energy}
+\keyword{internal}
 \keyword{tmp}
