+++ man/get_energy_service_transportation.Rd
@@ -13,3 +13,4 @@ energy_service_transportation_clean global variable
 Get transport.
 }
 \keyword{energy}
+\keyword{internal}
