+++ man/get_total_revenue.Rd
@@ -12,4 +12,5 @@ total_revenue global variable
 \description{
 Compute total revenue: total production * global price.
 }
+\keyword{internal}
 \keyword{revenue}
