+++ man/conv_EJ_GW.Rd
@@ -16,3 +16,5 @@ conv_EJ_GW(data, cf, EJ)
 \description{
 Covert EJ to GW
 }
+\keyword{conversion}
+\keyword{internal}
