+++ man/get_forcing.Rd
@@ -13,3 +13,4 @@ forcing_clean global variable
 Get World's forcing query.
 }
 \keyword{forcing}
+\keyword{internal}
