+++ man/get_co2_sequestration.Rd
@@ -13,3 +13,4 @@ co2_sequestration_clean global variable
 Get carbon sequestration.
 }
 \keyword{co2}
+\keyword{internal}
