+++ man/get_ag_prices_wld_tmp.Rd
@@ -13,4 +13,5 @@ ag_prices_wld global variable
 Get ag price index.
 }
 \keyword{ag}
+\keyword{internal}
 \keyword{tmp}
