+++ man/get_energy_trade.Rd
@@ -13,4 +13,5 @@ energy_trade_clean global variable
 Get energy trade supply.
 }
 \keyword{energy}
+\keyword{internal}
 \keyword{tmp}
