### 2023-12-21 67a2c1e
Review: fix non-ASCII characters (investment)
