### 2023-12-22 3e76088
Review: add vignette tutorial mapping modifications
