+++ vignettes/Modify_Mapping_Template_Tutorial.Rmd
@@ -1,0 +1,118 @@
+---
+title: "Modify Mappings and Template Tutorial"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Modify_Mapping_Template_Tutorial}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+In this tutorial, we will modify some of the pre-build mappings of the `gcamreport` package
+
+<span style="color:white">newline</span>
+
+
+
+
+
+
+
+
+
+## Example 1: step-by-step to add a new mapping item - *AvocadoSeeds* and *AvocadoSeedsTrees*
+
+In this example, we assume that the GCAM version has a more detailed land-food system and reports a new food item called *AvocadoSeeds* and a new land-leaf called *AvocadoSeedsTrees*. Thus, we need to modify `gcamreport` to include these items in the reporting dataset. In this example, the new items add to the final *cropland* area and several pollutant *emissions*, but do not have their own category in the template.
+
+1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R Studio or Docker.
+
+2) Open the `gcamreport.Rprj`.
+
+3) To modify the mappings, it is best to relay on some similar item that is already present in the mappings. In this case, we relay on *NutsSeeds* and *NutsSeedsTree*. To see where the mapping modifications should be performed, search across all files (`Ctrl + Shift + F`) inside `inst/extdata/mappings` for the key words *NutsSeeds* and *NutsSeedsTree*.
+   
+4) Using the previous point method, we can see that we should add some lines to the following files stored in `inst/extdata/mappings`:
+   
+   a. `ag_prices_map.csv`. Add this line: 
+   ```
+   AvocadoTrees,,,,,,,,,1
+   ```
+   b. `Kyotogas_sector.csv`. Add these lines: 
+   ```
+   CH4,AWB,AvocadoTrees,Emissions|Kyoto Gases,Emissions|Kyoto Gases|AFOLU,,,,,,,,,1
+   N2O,AGR,AvocadoTrees,Emissions|Kyoto Gases,Emissions|Kyoto Gases|AFOLU,,,,,,,,,1
+   N2O,AWB,AvocadoTrees,Emissions|Kyoto Gases,Emissions|Kyoto Gases|AFOLU,,,,,,,,,1
+   ```
+   c. `land_use_map.csv`. Add these line
+   ```
+   AvocadoTrees,Land Cover,Land Cover|Cropland,,,,,,,0.1
+   AvocadoTreesTree,Land Cover,Land Cover|Cropland,,,,,,,0.1
+   ```
+   d. `nonCO2_emissions_sector_map.csv`. Add these lines
+   ```
+   AvocadoTrees,BC_AWB,Emissions|BC,Emissions|BC|AFOLU,,,,,,,1
+   AvocadoTrees,CH4_AWB,Emissions|CH4,Emissions|CH4|AFOLU,,,,,,,1
+   AvocadoTrees,CO_AWB,Emissions|CO,Emissions|CO|AFOLU,,,,,,,1
+   AvocadoTrees,N2O_AGR,Emissions|N2O,Emissions|N2O|AFOLU,,,,,,,1
+   AvocadoTrees,N2O_AWB,Emissions|N2O,Emissions|N2O|AFOLU,,,,,,,1
+   AvocadoTrees,NH3_AGR,Emissions|NH3,Emissions|NH3|AFOLU,,,,,,,1
[…]
