### 2023-12-22 d8608f3
Add link to mapping modification vignette
