+++ paper/paper.md
@@ -84,9 +84,9 @@ The `gcamreport` package consists of a set of functions divided into two differe
 
 To read the raw GCAM outputs, `gcamreport` uses [`rgcam`](https://github.com/JGCRI/rgcam). This R package is part of the GCAM-ecosystem, which is a set of complementary tools to the GCAM model that extend its functionality and facilitate data management. In particular, `rgcam` allows to both read and create a manageable data file containing the desired model outputs selected from the GCAM results database. 
 
-For the correct performance of `gcamreport` internal functions, the package includes user-modifiable constant values and mapping files, such as regional and sectorial aggregations, as well as external files to compute additional calculations to produce some results that are beyond the regular GCAM reporting (e.g., installed capacity). These additional calculations are necessary to meet the IAMC reporting requirements (since not all models provide the same type of outputs, but often through back-on-the-envelope calculations, harmonizing outputs in this way improves the comparability of model results). Mapping files are automatically loaded and used, but they can be customized and adjusted by the user if required by the specific study or GCAM version as detailed in a specific [R vignette](TODO). In this case, it is recommended to save and tag the final used version to allow reusability and reproducibility.
+For the correct performance of `gcamreport` internal functions, the package includes user-modifiable constant values and mapping files, such as regional and sectorial aggregations, as well as external files to compute additional calculations to produce some results that are beyond the regular GCAM reporting (e.g., installed capacity). These additional calculations are necessary to meet the IAMC reporting requirements (since not all models provide the same type of outputs, but often through back-on-the-envelope calculations, harmonizing outputs in this way improves the comparability of model results). Mapping files are automatically loaded and used, but they can be customized and adjusted by the user if required by a specific study as described in a detailed [R vignette](https://bc3lc.github.io/gcamreport/articles/Modify_Mapping_Template_Tutorial.html). If the mapping files or the reporting template are modified, it is recommended to save and tag the final used version to allow reusability and reproducibility.
 
-The outputs generated by the functions within the *dataset generation block* consist of a RData dataset, a Comma-Separated Values (CSV) file, and a Microsoft Excel Open XML Spreadsheet (XLSX), all of which are automatically saved unless the user specifies otherwise with the `save_output` parameter. This ensures portability and format compatibility of the outputs [@krey_iamc_2023]. Besides, the outputs generated by the *interactive user block* are both CSV files and Portable Network Graphic (PNG) files, corresponding to the processed tabular data and the generated figures, which are only saved if the user clicks on the corresponding *download* button. This last block can be directly run with the RData dataset generated by the former block, avoiding to create again the reporting dataset as detailed in this [example](https://bc3lc.github.io/gcamreport/articles/Step_By_Step_Full_Example.html#example3).
+The output generated by the functions within the *dataset generation block* consists of an RData dataset, a Comma-Separated Values (CSV) file, and a Microsoft Excel Open XML Spreadsheet (XLSX), all of which are automatically saved unless the user specifies otherwise with the `save_output` parameter. This ensures portability and format compatibility of the output [@krey_iamc_2023]. Besides, the output generated by the *interactive user block* are both CSV files and Portable Network Graphic (PNG) files, corresponding to the processed tabular data and the generated figures, which are only saved if the user clicks on the corresponding *download* button. This last block can be directly run with the RData dataset generated by the former block, avoiding the need to re-create the reporting dataset, as detailed in this [example](https://bc3lc.github.io/gcamreport/articles/Step_By_Step_Full_Example.html#example3).
 
 (**a**)                                                          
 ![](gcamreport_table.png){width="100%"}
@@ -95,9 +95,9 @@ The outputs generated by the functions within the *dataset generation block* con
[…]
