+++ pkgdown.yml
@@ -6,12 +6,13 @@ articles:
   Dataset_Generation_Tutorial_v6-0: Dataset_Generation_Tutorial_v6-0.html
   Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
   Interactive_UI_Tutorial_v6-0: Interactive_UI_Tutorial_v6-0.html
+  Modify_Mapping_Template_Tutorial: Modify_Mapping_Template_Tutorial.html
   README_v6-0: README_v6-0.html
   Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
   Step_By_Step_Full_Example_v6-0: Step_By_Step_Full_Example_v6-0.html
   Version_Guide: Version_Guide.html
   gcamreport_v6-0: gcamreport_v6-0.html
-last_built: 2023-12-22T08:49Z
+last_built: 2023-12-22T09:21Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
