+++ R/main.R
@@ -14,24 +14,26 @@
 #' @importFrom xml2 read_xml xml_find_first
 #' @importFrom dplyr bind_rows
 #' @export
-data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regions = 'All') {
-  if (length(desired_regions) == 1 && desired_regions == 'All') {
+data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regions = "All") {
+  if (length(desired_regions) == 1 && desired_regions == "All") {
     desired_regions <- NULL
   }
 
   dt <- data.frame()
-  xml <- read_xml('inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml')
+  xml <- read_xml("inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml")
   qq <- xml_find_first(xml, paste0("//*[@title='", type, "']"))
 
   for (sc in scenarios) {
     emiss_list <- gcamreport::emissions_list
     while (length(emiss_list) > 0) {
-      current_emis <- emiss_list[1:min(21,length(emiss_list))]
+      current_emis <- emiss_list[1:min(21, length(emiss_list))]
       qq_sec <- gsub("current_emis", paste0("(@name = '", paste(current_emis, collapse = "' or @name = '"), "')"), qq)
 
       prj_tmp <- addSingleQuery(
         conn = localDBConn(db_path,
-                                  db_name,migabble = FALSE),
+          db_name,
+          migabble = FALSE
+        ),
         proj = prj_name,
         qn = type,
         query = qq_sec,
@@ -45,7 +47,7 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 
       tmp <- data.frame(prj_tmp[[sc]][type])
       if (nrow(tmp) > 0) {
-        dt <- bind_rows(dt,tmp)
+        dt <- bind_rows(dt, tmp)
       }
       rm(prj_tmp)
 
@@ -74,14 +76,14 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 #' @return loaded project into global environment
 #' @importFrom rgcam addSingleQuery localDBConn
 #' @export
-load_project <- function(project_path, desired_regions = 'All') {
-  print('Loading project...')
+load_project <- function(project_path, desired_regions = "All") {
+  print("Loading project...")
 
   # load the project
   prj <- loadProject(project_path)
 
   # filter the regions if not all of them are considered (desired_regions != 'All')
-  if (!(length(desired_regions) == 1 && desired_regions == 'All')) {
+  if (!(length(desired_regions) == 1 && desired_regions == "All")) {
     # for all scenarios in prj
     for (s in names(prj)) {
       # for all variables in prj
@@ -116,47 +118,54 @@ load_project <- function(project_path, desired_regions = 'All') {
 #' @import dplyr
 #' @export
 create_project <- function(db_path, db_name, prj_name, scenarios,
-                           desired_regions = 'All', desired_variables = 'All') {
+                           desired_regions = "All", desired_variables = "All") {
   Internal_variable <- Variable <- required <- NULL
 
   # check if the project already exists
-  file_name <- file.path(db_path, paste(db_name, prj_name, sep = '_'))
+  file_name <- file.path(db_path, paste(db_name, prj_name, sep = "_"))
   if (file.exists(file_name)) {
     load_project(file_name, desired_regions)
-
   } else {
-    print('Creating project...')
+    print("Creating project...")
 
     # create the project
     conn <- localDBConn(db_path,
-                        db_name,migabble = FALSE)
+      db_name,
+      migabble = FALSE
+    )
[…]
