+++ man/land_use_map.Rd
@@ -18,8 +18,8 @@ ag maps
 }
 \examples{
 \dontrun{
- library(gcamreport);
- gcamreport::land_use_map
+library(gcamreport)
+gcamreport::land_use_map
 }
 }
 \keyword{datasets}
