+++ man/GHG_gases.Rd
@@ -15,8 +15,8 @@ ghg emission conversion
 }
 \examples{
 \dontrun{
- library(gcamreport);
- gcamreport::GHG_gases
+library(gcamreport)
+gcamreport::GHG_gases
 }
 }
 \keyword{datasets}

[…]
