### 2023-12-22 #5975 @klau506
Dear @bpbond and @martinfleis,

Thank you for the detailed review and the constructive comments. The package has gained robustness and usability while being cleaned and optimized.

First, I would like to note that during the review process I have updated the package to meet the novelties of GCAM 7. The new version has updated the mappings and improved some minor issues found during these weeks: now, it is possible to create the reporting dataset for pre-defined variables and/or regions. Moreover, the package has more input checks, improves and includes new documentation pages, and fixes some internal minor bugs in some calculations. Some functions have also been renamed to make them more intuitive, and the usage of some functions have been simplified optimizing the code (Tagging also @ibarraespinosa).

Going through the specific issues from reviewer @bpbond, the following changes have been made:

- In the manuscript (_[Manuscript suggestions](https://github.com/bc3LC/gcamreport/issues/4)_ issue):
     - With the help of comments 1-10, the manuscript text has been improved.
     - Following comment 5, the manuscript refers to a [step-by-step tutorial](https://bc3lc.github.io/gcamreport/articles/Step_By_Step_Full_Example.html) and gives a short example of how to use the package.
     - Following comment 11, the manuscript refers to a [step-by-step tutorial](https://bc3lc.github.io/gcamreport/articles/Modify_Mapping_Template_Tutorial.html) on modifying the mapping files and reporting template. Due to its length and technical detail, we decided to include this description as an R vignette rather than an explanation embedded in the text. This vignette can be directly opened in the manuscript in line 100.
     - Following comment 12, Figure 2 has been enlarged to improve readability.
     - Following comment 13, a paragraph about performance has been added (line 118). It details the package running time and how it depends on the project creation time relaying on the `rgcam` package.
- In the code (_[Code suggestions](https://github.com/bc3LC/gcamreport/issues/5)_ issue):
     - Following comment 1, `paste0` has been replaced with `file.path` to increase robustness.
     - Following comment 2, `=` assignments have been replaced with for `<-` as recommended.
     - Following comment 3, the package has been run through the `styler` package to ensure clear and consistent code style.
     - Following comment 4, the `rm(list = ls())` lines have been removed from the test blocks.
     - Following comment 5, the `library` calls have been removed from the code.
     - Following comment 6, the `importFrom` function has been added to the code, deleting the `dplyr::`-type namespace quantifiers to improve the readability and broaden the usage of the `Roxygen` package.
     - Following comment 7, non-ASCII characters have been re-encoded.
     - Following comment 7, errors, warnings, and notes have been reduced when using the `R CMD Check`. There are currently no errors or warnings, and only four notes.
     - Following comment 8, some internal warnings have been removed to clean up the user output and improve the readability of the test results. However, the `rgcam` package, which produces the project file, has a lot of printed output that could not be removed. For instance,
```r
About to perform query: collection()/scenario
About to perform query: collection()/scenario[ (@name='Reference' and @date='2023-4-12T16:34:34+01:00') ]/world/region/nationalAccount/account[@name = 'GDP']/node()
In Function: 1703243758723
After Function: 1703243758832
```

Thank you for your time and hope that the performed modifications are correct.

Best wishes,
Clàudia
