### 2024-01-03 a2070c1
Allow default selection of all db/prj scenarios. Improve user's input checks. Add tests
