+++ tests/testthat/testReporter_v7.0.R
@@ -186,56 +186,84 @@ test_that("Test7_v7. specify variables, regions, continents", {
 test_that("Test8_v7. error messages", {
 
   expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
-                               db_name = 'database_basexdb',
+                               db_name = 'database_basexdb_ref',
                                prj_name = 'gcamv7.0_noCreated.dat',
                                scenarios = 'Reference',
                                desired_variables = 'dummy variable',
                                launch_ui = FALSE),
-               "ERROR: You specified the variable dummy variable which is not available for reporting.")
+               "The variable dummy variable is not available for reporting")
+  expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
+                               db_name = 'database_basexdb_ref',
+                               prj_name = 'gcamv7.0_noCreated.dat',
+                               scenarios = 'Reference',
+                               desired_variables = c('dummy1', 'dummy2'),
+                               launch_ui = FALSE),
+               "The variables dummy1, dummy2 are not available for reporting")
 
   expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
-                               db_name = 'database_basexdb',
+                               db_name = 'database_basexdb_ref',
                                prj_name = 'gcamv7.0_noCreated.dat',
                                scenarios = 'Reference',
                                desired_variables = 'Final|Energy*',
                                launch_ui = FALSE),
-               "ERROR: You specified the variable Final|Energy* which is not available for reporting.")
-
+               "There is no variable containing the pattern Final|Energy* available for reporting.")
   expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
-                               db_name = 'database_basexdb',
+                               db_name = 'database_basexdb_ref',
                                prj_name = 'gcamv7.0_noCreated.dat',
                                scenarios = 'Reference',
-                               desired_region = 'dummy region',
+                               desired_variables = c('Final|Energy*','Emissions CH4*'),
                                launch_ui = FALSE),
-               "ERROR: You specified the region dummy region which is not available for reporting.")
+               "There are no variables containing the patterns Final|Energy*, Emissions CH4* available for reporting")
 
   expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
-                               db_name = 'database_basexdb',
+                               db_name = 'database_basexdb_ref',
+                               prj_name = 'gcamv7.0_noCreated.dat',
+                               scenarios = 'Reference',
+                               desired_regions = 'dummy region',
+                               launch_ui = FALSE),
+               "The desired region dummy region is not available for reporting.")
+  expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
+                               db_name = 'database_basexdb_ref',
+                               prj_name = 'gcamv7.0_noCreated.dat',
+                               scenarios = 'Reference',
+                               desired_regions = c('dummy1', 'dummy2'),
+                               launch_ui = FALSE),
+               "The desired regions dummy1, dummy2 are not available for reporting.")
+
+  expect_error(generate_report(db_path = file.path(rprojroot::find_root(rprojroot::is_testthat),'testInputs/v_7.0/'),
[…]
