### 2024-01-03 e5eba6c
Allow default selection of all db/prj scenarios. Improve user's input checks. Add tests
