+++ R/main.R
@@ -1,21 +1,21 @@
 #' data_query
 #'
 #' Add nonCO2 large queries
-#' @param db_path path of the database
-#' @param db_name name of the database
-#' @param prj_name name of the project
-#' @param scenarios name of the scenarios to be considered
-#' @param type either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'
+#' @param db_path path of the database.
+#' @param db_name name of the database.
+#' @param prj_name name of the project.
+#' @param scenarios name of the scenarios to be considered.
+#' @param type either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'.
 #' @param desired_regions desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
 #' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
-#' @return dataframe with the data from the query
+#' @return dataframe with the data from the query.
 #' @importFrom rgcam addSingleQuery localDBConn
 #' @importFrom xml2 read_xml xml_find_first
 #' @importFrom dplyr bind_rows
 #' @export
 data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regions = "All") {
-  if (length(desired_regions) == 1 && desired_regions == "All") {
+  if (identical(desired_regions, "All")) {
     desired_regions <- NULL
   }
 
@@ -69,21 +69,23 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 #' load_project
 #'
 #' Load specified project into the global environment
-#' @param project_path path of the project (including project name and extension)
+#' @param project_path path of the project (including project name and extension).
 #' @param desired_regions desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
 #' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
-#' @return loaded project into global environment
-#' @importFrom rgcam addSingleQuery localDBConn
+#' @param scenarios name of the scenarios to be considered. By default, all the scenarios in the project are considered.
+#' @return loaded project into global environment.
+#' @importFrom rgcam addSingleQuery localDBConn listScenarios loadProject dropScenarios
+#' @importFrom dplyr intersect setdiff if_else
 #' @export
-load_project <- function(project_path, desired_regions = "All") {
+load_project <- function(project_path, desired_regions = "All", scenarios = NULL) {
   print("Loading project...")
 
   # load the project
   prj <- loadProject(project_path)
 
   # filter the regions if not all of them are considered (desired_regions != 'All')
-  if (!(length(desired_regions) == 1 && desired_regions == "All")) {
+  if (!(identical(desired_regions, "All"))) {
     # for all scenarios in prj
     for (s in names(prj)) {
       # for all variables in prj
@@ -92,7 +94,22 @@ load_project <- function(project_path, desired_regions = "All") {
       }
     }
   }
-  scenarios.global <<- listScenarios(prj)
+  if (is.null(scenarios)) {
+    scenarios.global <<- listScenarios(prj)
+  } else {
+    scenarios.global <<- intersect(scenarios,listScenarios(prj))
+    # check the user input
+    if (length(scenarios) > length(scenarios.global)) {
+      check_scen <- setdiff(scenarios,listScenarios(prj))
+      tmp <- paste(check_scen, collapse = ', ')
+      if (length(check_scen) > 1) stop("The desired scenarios ",tmp, " are not present in the loaded project.\n")
[…]
