+++ man/load_query.Rd
@@ -7,14 +7,14 @@
 load_query(var, base_data, final_queries)
 }
 \arguments{
-\item{var}{variable to be loaded}
+\item{var}{variable to be loaded.}
 
-\item{base_data}{dataframe with the required internal variables}
+\item{base_data}{dataframe with the required internal variables.}
 
-\item{final_queries}{vector of the queries to be loaded}
+\item{final_queries}{vector of the queries to be loaded.}
 }
 \value{
-query name to be loaded
+query name to be loaded.
 }
 \description{
 Recursive function to load the necessary queries for the desired variables
