+++ man/data_query.Rd
@@ -14,22 +14,22 @@ data_query(
 )
 }
 \arguments{
-\item{type}{either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'}
+\item{type}{either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'.}
 
-\item{db_path}{path of the database}
+\item{db_path}{path of the database.}
 
-\item{db_name}{name of the database}
+\item{db_name}{name of the database.}
 
-\item{prj_name}{name of the project}
+\item{prj_name}{name of the project.}
 
-\item{scenarios}{name of the scenarios to be considered}
+\item{scenarios}{name of the scenarios to be considered.}
 
 \item{desired_regions}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
 To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 In case the project dataset needs to be created, it will be produced with only the specified regions.}
 }
 \value{
-dataframe with the data from the query
+dataframe with the data from the query.
 }
 \description{
 Add nonCO2 large queries
