+++ inst/gcamreport_ui/ui.R
@@ -31,7 +31,7 @@ ui <- dashboardPage(
           choices = unique(sdata$Scenario),
           selected = unique(sdata$Scenario)
         ) %>%
-          tagAppendAttributes(class = 'names_split')
+          tagAppendAttributes(class = "names_split")
       ),
       # select none/all buttons
       menuItem(
@@ -39,7 +39,7 @@ ui <- dashboardPage(
           inputId = "select_all_scen",
           label = "Select all",
           style = "minimal",
-          size = 'xs'
+          size = "xs"
         )
       ),
       menuItem(
@@ -47,7 +47,7 @@ ui <- dashboardPage(
           inputId = "select_none_scen",
           label = "Select none",
           style = "minimal",
-          size = 'xs'
+          size = "xs"
         )
       ),
       br()
@@ -61,8 +61,9 @@ ui <- dashboardPage(
       startExpanded = FALSE,
       menuItem(
         shinyTree("tree_regions",
-                  checkbox = TRUE,
-                  types = "{ 'basic': {'a_attr' : { 'style' : 'background-color: #2c3b41; color: inherit; cursor: inherit; pointer-events: inherit; opacity: inherit'}}}")
+          checkbox = TRUE,
+          types = "{ 'basic': {'a_attr' : { 'style' : 'background-color: #2c3b41; color: inherit; cursor: inherit; pointer-events: inherit; opacity: inherit'}}}"
+        )
       ),
       # select none/all buttons
       menuItem(
@@ -70,7 +71,7 @@ ui <- dashboardPage(
           inputId = "select_all_regions",
           label = "Select all",
           style = "minimal",
-          size = 'xs'
+          size = "xs"
         )
       ),
       menuItem(
@@ -78,7 +79,7 @@ ui <- dashboardPage(
           inputId = "select_none_regions",
           label = "Select none",
           style = "minimal",
-          size = 'xs'
+          size = "xs"
         )
       ),
       br()
@@ -92,9 +93,10 @@ ui <- dashboardPage(
       startExpanded = FALSE,
       menuItem(
         shinyTree("tree_variables",
-                  checkbox = TRUE,
-                  types = "{ 'basic': {'a_attr' : { 'style' : 'background-color: #2c3b41; color: inherit; cursor: inherit; pointer-events: inherit; opacity: inherit'}},
-                  'dis': {'a_attr' : { 'style' : 'opacity: 0.3; cursor: not-allowed; pointer-events: none' } }}")
+          checkbox = TRUE,
+          types = "{ 'basic': {'a_attr' : { 'style' : 'background-color: #2c3b41; color: inherit; cursor: inherit; pointer-events: inherit; opacity: inherit'}},
+                  'dis': {'a_attr' : { 'style' : 'opacity: 0.3; cursor: not-allowed; pointer-events: none' } }}"
+        )
       ),
       # select none/all buttons
       menuItem(
@@ -102,7 +104,7 @@ ui <- dashboardPage(
           inputId = "select_all_variables",
           label = "Select all",
           style = "minimal",
-          size = 'xs'
+          size = "xs"
         )
       ),
       menuItem(
@@ -110,7 +112,7 @@ ui <- dashboardPage(
           inputId = "select_none_variables",
           label = "Select none",
           style = "minimal",
-          size = 'xs'
+          size = "xs"
         )
       ),
       br()
@@ -129,7 +131,7 @@ ui <- dashboardPage(
           choices = available_years,
           selected = available_years
         ) %>%
-          tagAppendAttributes(class = 'names_split')
+          tagAppendAttributes(class = "names_split")
       ),
       # select none/all buttons
       menuItem(
@@ -137,7 +139,7 @@ ui <- dashboardPage(
[…]

[…]
