### 2024-01-03 44a4127
Update vignettes with scenarios novelties
