### 2024-01-10 #5975 @klau506
Dear @martinfleis,

Thank you for managing the manuscript submission. During this time, we have made some additional improvements to the package in order to increase the robustness in collaboration with some researchers from the GCAM team. Since the review is still ongoing, and due to his relevant contribution, we would like to add a new co-author (Matthew Binsted) to the paper. Just wanted to confirm if is possible. Let us know otherwise.

Thank you so much for your time and consideration.
Looking forward to address any new comment/suggestion that comes up in the review,
Clàudia
