### 2024-01-17 #5975 @ibarraespinosa
Sure. I will finish heliio this week and I will start gcamrepprt

On Wed, Jan 17, 2024, 3:36 AM Martin Fleischmann ***@***.***>
wrote:

> Hi @ibarraespinosa <https://github.com/ibarraespinosa>, can you give us
> an update on the progress of your review? Thanks!
>
