### 2024-01-22 4215600
Fix Capital Cost mapping
