+++ R/main.R
@@ -1,37 +1,36 @@
-## Main functions of the package
-library(usethis)
-library(magrittr)
-
 #' data_query
 #'
 #' Add nonCO2 large queries
-#' @param db_path: path of the database
-#' @param db_name: name of the database
-#' @param prj_name: name of the project
-#' @param scenarios: name of the scenarios to be considered
-#' @param type: either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'
-#' @param desired_regions: desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
+#' @param db_path path of the database
+#' @param db_name name of the database
+#' @param prj_name name of the project
+#' @param scenarios name of the scenarios to be considered
+#' @param type either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'
+#' @param desired_regions desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
 #' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
 #' @return dataframe with the data from the query
+#' @importFrom rgcam addSingleQuery localDBConn
+#' @importFrom xml2 read_xml xml_find_first
+#' @importFrom dplyr bind_rows
 #' @export
-data_query = function(type, db_path, db_name, prj_name, scenarios, desired_regions = 'All') {
+data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regions = 'All') {
   if (length(desired_regions) == 1 && desired_regions == 'All') {
-    desired_regions = NULL
+    desired_regions <- NULL
   }
 
-  dt = data.frame()
-  xml <- xml2::read_xml('inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml')
-  qq <- xml2::xml_find_first(xml, paste0("//*[@title='", type, "']"))
+  dt <- data.frame()
+  xml <- read_xml('inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml')
+  qq <- xml_find_first(xml, paste0("//*[@title='", type, "']"))
 
   for (sc in scenarios) {
-    emiss_list = emissions_list
+    emiss_list <- emissions_list
     while (length(emiss_list) > 0) {
-      current_emis = emiss_list[1:min(21,length(emiss_list))]
-      qq_sec = gsub("current_emis", paste0("(@name = '", paste(current_emis, collapse = "' or @name = '"), "')"), qq)
+      current_emis <- emiss_list[1:min(21,length(emiss_list))]
+      qq_sec <- gsub("current_emis", paste0("(@name = '", paste(current_emis, collapse = "' or @name = '"), "')"), qq)
 
-      prj_tmp = rgcam::addSingleQuery(
-        conn = rgcam::localDBConn(db_path,
+      prj_tmp <- addSingleQuery(
+        conn = localDBConn(db_path,
                                   db_name,migabble = FALSE),
         proj = prj_name,
         qn = type,
@@ -44,16 +43,16 @@ data_query = function(type, db_path, db_name, prj_name, scenarios, desired_regio
         warn.empty = FALSE
       )
 
-      tmp = data.frame(prj_tmp[[sc]][type])
+      tmp <- data.frame(prj_tmp[[sc]][type])
       if (nrow(tmp) > 0) {
-        dt = dplyr::bind_rows(dt,tmp)
+        dt <- bind_rows(dt,tmp)
       }
       rm(prj_tmp)
 
       if (length(emiss_list) > 21) {
         emiss_list <- emiss_list[(21 + 1):length(emiss_list)]
       } else {
-        emiss_list = c()
[…]
