+++ inst/extdata/saveDataFiles.R
@@ -1,7 +1,6 @@
 # Converting raw data into package data
 library(usethis)
 library(magrittr)
-library(here)
 
 ### paths
 rawDataFolder = here::here()
@@ -16,12 +15,12 @@ emissions_list = c('BC','BC_AWB','C2F6','CF4','CH4','CH4_AGR','CH4_AWB','CO','CO
 use_data(emissions_list, overwrite=T)
 
 # regions_continents_map
-reg_cont <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/regions_continents_map.csv"), skip = 1,
+reg_cont <- read.csv(file.path(rawDataFolder, "inst/extdata/mappings", "regions_continents_map.csv"), skip = 1,
                      stringsAsFactors = FALSE)
 use_data(reg_cont, overwrite=T)
 
 # variables_functions_mapping
-var_fun_map <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/variables_functions_mapping.csv"),
+var_fun_map <- read.csv(file.path(rawDataFolder, "inst/extdata/mappings", "variables_functions_mapping.csv"),
                        sep=';',header=T, na.strings=c("","NA"), stringsAsFactors = FALSE)
 
 var_fun_map$dependencies <- as.list(strsplit(var_fun_map$dependencies, ","))
@@ -30,127 +29,130 @@ var_fun_map$queries <- as.list(strsplit(var_fun_map$queries, ","))
 use_data(var_fun_map, overwrite=T)
 
 # ghg adjuster
-GWP_adjuster <- read.csv(paste0(here::here(), "/inst/extdata/mappings", "/ghg_GWP.csv"), skip = 1, na = "",
+GWP_adjuster <- read.csv(file.path(rawDataFolder, "inst/extdata/mappings", "ghg_GWP.csv"), skip = 1, na = "",
                          stringsAsFactors = FALSE)
 use_data(GWP_adjuster, overwrite=T)
 
 # vetting test
-global_vet_values <- read.csv(paste0(here::here(), "/inst/extdata/vetting", "/global_vet_values.csv"),
+global_vet_values <- read.csv(file.path(rawDataFolder, "inst/extdata/vetting", "global_vet_values.csv"),
                               stringsAsFactors = FALSE)
 use_data(global_vet_values, overwrite=T)
 
 # Read in template
-template <- read.csv(paste0(rawDataFolder, "/inst/extdata", "/template/reporting_template.csv"),
+template <- read.csv(file.path(rawDataFolder, "inst/extdata", "template/reporting_template.csv"),
                      fileEncoding = "UTF-8-BOM", stringsAsFactors = FALSE)
 use_data(template, overwrite=T)
 
 # emissions maps
-co2_sector_map <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/CO2_sector_map.csv"), skip = 1, na = "",
+co2_sector_map <- read.csv(file.path(rawDataFolder, "inst/extdata/mappings", "CO2_sector_map.csv"), skip = 1, na = "",
                            stringsAsFactors = FALSE) %>% gather_map()
 use_data(co2_sector_map, overwrite=T)
 
-co2_ets_sector_map <- read.csv(paste0(rawDataFolder, "/inst/extdata/mappings", "/CO2_ETS_sector_map.csv"), skip = 1, na = "",
+co2_ets_sector_map <- read.csv(file.path(rawDataFolder, "inst/extdata/mappings", "CO2_ETS_sector_map.csv"), skip = 1, na = "",
                            stringsAsFactors = FALSE) %>% gather_map()
 use_data(co2_ets_sector_map, overwrite=T)
 
[…]

[…]
