+++ R/app_functions.R
@@ -11,8 +11,8 @@
 #' @importFrom dplyr if_else
 #' @export
 compute_height <- function(reg_all) {
-  reg_clean <- setdiff(reg_all, c('MAF','LAM','OECD90','REF','ASIA'))
-  h <- if_else(length(reg_clean) < 3, 325, 160*ceiling(sqrt(length(reg_clean))))
+  reg_clean <- setdiff(reg_all, c("MAF", "LAM", "OECD90", "REF", "ASIA"))
+  h <- if_else(length(reg_clean) < 3, 325, 160 * ceiling(sqrt(length(reg_clean))))
 
   return(h)
 }
@@ -25,8 +25,8 @@ compute_height <- function(reg_all) {
 #' @keywords internal
 #' @export
 is_leaf <- function(tree) {
-  if (length(tree) > 1 || length(tree[[1]]) > 1 || length(tree[[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]]) > 1
-      || length(tree[[1]][[1]][[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]][[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]][[1]][[1]][[1]]) > 1) {
+  if (length(tree) > 1 || length(tree[[1]]) > 1 || length(tree[[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]]) > 1 ||
+      length(tree[[1]][[1]][[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]][[1]][[1]]) > 1 || length(tree[[1]][[1]][[1]][[1]][[1]][[1]]) > 1) {
     return(FALSE)
   }
 
@@ -44,7 +44,7 @@ is_leaf <- function(tree) {
 #' @importFrom magrittr %>%
 #' @export
 change_style <- function(tree, type, tmp_vars = NULL) {
-  n = length(tree)
+  n <- length(tree)
   for (i in 1:n) {
     # get the attributes for the current node
     attrs <- attributes(tree[[i]])
@@ -57,10 +57,10 @@ change_style <- function(tree, type, tmp_vars = NULL) {
 
     # change the style for the current node. If it's in the list of variables to be disabled, set
     # it to 'dis'. To 'basic' otherwise
-    if (type == 'variables' && length(tmp_vars) > 0 && attrs$my_id %in% tmp_vars) {
-      attrs$sttype <- 'dis'
+    if (type == "variables" && length(tmp_vars) > 0 && attrs$my_id %in% tmp_vars) {
+      attrs$sttype <- "dis"
     } else {
-      attrs$sttype <- 'basic'
+      attrs$sttype <- "basic"
     }
     attributes(tree[[i]]) <- attrs
   }
@@ -83,7 +83,6 @@ change_style <- function(tree, type, tmp_vars = NULL) {
 #' @importFrom stringr str_extract
 #' @export
 check_user_choices_plot <- function(vars, scen, years, reg, grouped) {
-
   # errors' vector
   error_message <- c()
 
@@ -96,23 +95,23 @@ check_user_choices_plot <- function(vars, scen, years, reg, grouped) {
 
   # check that at least one scenario has been choosen
   if (length(unique(scen)) < 1) {
-    error_message <- c(error_message,"ERROR: Select at least one scenario please.")
+    error_message <- c(error_message, "ERROR: Select at least one scenario please.")
   }
   # check that at least one year has been choosen
   if (length(unique(years)) < 1) {
-    error_message <- c(error_message,"ERROR: Select at least one year please.")
+    error_message <- c(error_message, "ERROR: Select at least one year please.")
   }
   # check that at least one region has been choosen
   if (length(unique(reg)) < 1) {
[…]
