+++ man/get_energy_trade_tmp.Rd
@@ -13,9 +13,5 @@ energy_trade_supply global variable
 Get energy trade supply. Query used to compute other variables.
 }
 \keyword{energy}
-<<<<<<< HEAD
-\keyword{rgcam::getQuery}
-=======
 \keyword{internal}
->>>>>>> dc7f961 (Review: fix warning message undefined data)
 \keyword{tmp}
