+++ man/co2_market_frag_map.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{co2_market_frag_map}
+\alias{co2_market_frag_map}
+\title{co2_market_frag_map}
+\format{
+.csv
+}
+\source{
+github
+}
+\usage{
+co2_market_frag_map
+}
+\description{
+new CO2 market fragmented map
+}
+\examples{
+\dontrun{
+ library(gcamreport);
+ gcamreport::co2_market_frag_map
+}
+}
+\keyword{datasets}

[…]
