+++ R/main.R
@@ -20,18 +20,20 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
   }
 
   dt <- data.frame()
-  xml <- read_xml('inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml')
+  xml <- read_xml("inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml")
   qq <- xml_find_first(xml, paste0("//*[@title='", type, "']"))
 
   for (sc in scenarios) {
     emiss_list <- gcamreport::emissions_list
     while (length(emiss_list) > 0) {
-      current_emis <- emiss_list[1:min(21,length(emiss_list))]
+      current_emis <- emiss_list[1:min(21, length(emiss_list))]
       qq_sec <- gsub("current_emis", paste0("(@name = '", paste(current_emis, collapse = "' or @name = '"), "')"), qq)
 
       prj_tmp <- addSingleQuery(
         conn = localDBConn(db_path,
-                                  db_name,migabble = FALSE),
+          db_name,
+          migabble = FALSE
+        ),
         proj = prj_name,
         qn = type,
         query = qq_sec,
@@ -45,7 +47,7 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 
       tmp <- data.frame(prj_tmp[[sc]][type])
       if (nrow(tmp) > 0) {
-        dt <- bind_rows(dt,tmp)
+        dt <- bind_rows(dt, tmp)
       }
       rm(prj_tmp)
 
@@ -95,13 +97,13 @@ load_project <- function(project_path, desired_regions = "All", scenarios = NULL
   if (is.null(scenarios)) {
     scenarios.global <<- listScenarios(prj)
   } else {
-    scenarios.global <<- intersect(scenarios,listScenarios(prj))
+    scenarios.global <<- intersect(scenarios, listScenarios(prj))
     # check the user input
     if (length(scenarios) > length(scenarios.global)) {
-      check_scen <- setdiff(scenarios,listScenarios(prj))
-      tmp <- paste(check_scen, collapse = ', ')
-      if (length(check_scen) > 1) stop("The desired scenarios ",tmp, " are not present in the loaded project.\n")
-      if (length(check_scen) == 1) stop("The desired scenario ",tmp, " is not present in the loaded project.\n")
+      check_scen <- setdiff(scenarios, listScenarios(prj))
+      tmp <- paste(check_scen, collapse = ", ")
+      if (length(check_scen) > 1) stop("The desired scenarios ", tmp, " are not present in the loaded project.\n")
+      if (length(check_scen) == 1) stop("The desired scenario ", tmp, " is not present in the loaded project.\n")
     }
     # drop unnecessary scenarios
     for (i in listScenarios(prj)[!listScenarios(prj) %in% scenarios]) {
@@ -137,15 +139,17 @@ create_project <- function(db_path, db_name, prj_name, scenarios,
   Internal_variable <- Variable <- required <- available_scenarios <- name <- NULL
 
   # check if the project already exists
-  file_name <- file.path(db_path, paste(db_name, prj_name, sep = '_'))
+  file_name <- file.path(db_path, paste(db_name, prj_name, sep = "_"))
   if (file.exists(file_name)) {
     load_project(file_name, desired_regions, scenarios)
   } else {
-    print('Creating project...')
+    print("Creating project...")
 
     # create the project
     conn <- localDBConn(db_path,
-                        db_name,migabble = FALSE)
+      db_name,
+      migabble = FALSE
+    )
 
     available_scenarios <- listScenariosInDB(conn) %>%
       pull(name)
[…]
