### 2024-02-19 bc3LC/gcamreport#6 @ibarraespinosa
Review for JOSS
Hi,

https://github.com/openjournals/joss-reviews/issues/5975

I'm sorry it took me ages to start review the repo and manuscript. Hopefully, as already was reviewewd once, it will be a smooth process.

I will be comenting here all my doubts and questions and once they al solved or clarified it is done.
