### 2024-02-19 bc3LC/gcamreport#6 @ibarraespinosa
# first example does not work

```

> # examples
> ## -- load gcamreport library.
> library(gcamreport)
> ## -- store the database path, name, and scenarios in a variable.
> dbpath <- "examples"
> dbname <- "database_basexdb_ref"
> scen <- "Reference"
> ## -- choose a project name
> prjname <- "example1.dat"
> ## -- generate the reporting dataset until 2050 for EU-12 and EU-15 for all the
> ## -- Agricultural variables, save the output in .RData, .csv and .xlsx format,
> ## -- and lunch the user interface
> generate_report(db_path = dbpath, db_name = dbname, scenarios = scen,
+                 prj_name = prjname, final_year = 2050,
+                 desired_regions = c('EU-12', 'EU-15'),
+                 desired_variables = c('Agricultural*'),
+                 save_output = TRUE, launch_ui = TRUE)
[1] "Creating project..."
/home/sergio/basex/.basex: writing new configuration file.
Error in localDBConn(db_path, db_name, migabble = FALSE) :
  Database does not exist or is invalid: examples/database_basexdb_ref
In addition: Warning messages:
1: In normalizePath(dbPath) :
  path[1]="examples": No such file or directory
2: The following named parsers don't match the column names: name, date, version

```
