### 2024-02-19 bc3LC/gcamreport#6 @ibarraespinosa
shiny needs to be added in imports or depends in DESCRIPTION
