### 2024-02-19 bc3LC/gcamreport#6 @ibarraespinosa
launch_gcamreport_ui("examples/example3.RData")

Listening on http://127.0.0.1:7278

Attaching package: ‘magrittr’

The following objects are masked from ‘package:testthat’:

    equals, is_less_than, not

Attaching package: ‘shinydashboard’

The following object is masked from ‘package:graphics’:

    box

Learn different usages for shinyjs and other Shiny tricks:
	https://deanattali.com/blog/advanced-shiny-tips

Attaching package: ‘shinyjs’

The following object is masked from ‘package:shinyWidgets’:

    alert

The following object is masked from ‘package:shiny’:

    runExample

The following objects are masked from ‘package:methods’:

    removeClass, show

Warning in file(con, "r") :
  cannot open file '/home/sergio/R/www/style.css': No such file or directory
Warning: Error in file: cannot open the connection
  72: file
  71: readLines
  70: includeCSS
   2: runApp
   1: launch_gcamreport_ui [/home/sergio/gcamreport/R/main.R#724]
