### 2024-02-19 #5975 @ibarraespinosa
reviewing here https://github.com/bc3LC/gcamreport/issues/6
