+++ man/gcamreport.Rd
@@ -1,6 +1,6 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/gcamreport.R
-\docType{package}
+\docType{_PACKAGE}
 \name{gcamreport}
 \alias{gcamreport}
 \title{gcamreport}
