+++ tests/testthat/testReporter_v7.0.R
@@ -21,7 +21,7 @@ test_that("Test1_v7. download db, create project, and run", {
   testResult <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test7.dat")))
   prj <- prj_tmp
   testthat::expect_equal(prj$Reference$`nonCO2 emissions by region`, testResult$Reference$`nonCO2 emissions by region`)
-  testthat::expect_equal(prj$Reference$`nonCO2 emissions by sector`, testResult$Reference$`nonCO2 emissions by sector (excluding resource production)`)
+  testthat::expect_equal(prj$Reference$`nonCO2 emissions by sector (excluding resource production)`, testResult$Reference$`nonCO2 emissions by sector (excluding resource production)`)
   testthat::expect_equal(prj$Reference$`CO2 prices`, testResult$Reference$`CO2 prices`)
 
   # check nonCO2 emissions query
