+++ pkgdown.yml
@@ -4,6 +4,7 @@ pkgdown_sha: ~
 articles:
   Dataset_Generation_Tutorial: Dataset_Generation_Tutorial.html
   Dataset_Generation_Tutorial_v6-0: Dataset_Generation_Tutorial_v6-0.html
+  gcamreport_v6-0: gcamreport_v6-0.html
   Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
   Interactive_UI_Tutorial_v6-0: Interactive_UI_Tutorial_v6-0.html
   Modify_Mapping_Template_Tutorial: Modify_Mapping_Template_Tutorial.html
@@ -11,8 +12,7 @@ articles:
   Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
   Step_By_Step_Full_Example_v6-0: Step_By_Step_Full_Example_v6-0.html
   Version_Guide: Version_Guide.html
-  gcamreport_v6-0: gcamreport_v6-0.html
-last_built: 2024-02-21T14:20Z
+last_built: 2024-02-21T15:02Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
