+++ vignettes/Interactive_UI_Tutorial.Rmd
@@ -41,7 +41,7 @@ After generating an `.RData` dataset with the `generate_report()` function, simp
 # lanuch the user interface specifying the dataset variable name
 launch_gcamreport_ui(data = "report") # report is the name of the standardized 
                                       # dataset generated by generate_report() 
-                                      # function. # So, if you see it loaded in 
+                                      # function. So, if you see it loaded in 
                                       # your environment, you can easily use 
                                       # it here!! :)
 
@@ -56,9 +56,77 @@ launch_gcamreport_ui(data_path = /path/to/your/database/..._standardize.RData)
 
 
 
-## Troubleshooting when launching the UI with the Docker installation
+## Troubleshooting when launching the UI
 
-#### A) Wired message when launching the UI when using the Docker installation.
+#### A) Error when launching the UI
+
+When running `launch_gcamreport_ui(data_path = "path/to/your/data.RData")`, you might get this message:
+
+
+    launch_gcamreport_ui(data_path = "examples/example3.RData")
+
+    Listening on http://127.0.0.1:7278
+
+    Attaching package: ‘magrittr’
+
+    The following objects are masked from ‘package:testthat’:
+
+    equals, is_less_than, not
+    Attaching package: ‘shinydashboard’
+
+    The following object is masked from ‘package:graphics’:
+
+    box
+    Learn different usages for shinyjs and other Shiny tricks:
+    https://deanattali.com/blog/advanced-shiny-tips
+
+    Attaching package: ‘shinyjs’
+
+    The following object is masked from ‘package:shinyWidgets’:
+
+    alert
+    The following object is masked from ‘package:shiny’:
+
+    runExample
+    The following objects are masked from ‘package:methods’:
+
+    removeClass, show
+    Warning in file(con, "r") :
+    cannot open file '/home/user/R/www/style.css': No such file or directory
+    Warning: Error in file: cannot open the connection
+    72: file
+    71: readLines
+    70: includeCSS
+    2: runApp
+    1: launch_gcamreport_ui [/home/user/gcamreport/R/main.R#724]
+
+**Possible solution 1**: make sure that you cloned the repo. Check the instructions [here](https://bc3lc.github.io/gcamreport/index.html#installation-guide).
+
+**Possible solution 2**: make sure that you are using the `gcamreport` Rproject and that the R session working directory points the `gcamreport` cloned folder. In addition:
+
+-   In case you are using the `gcamreport` package following the [R installation](#with-r), try to copy the whole path to your data, for instance `data_path = C:\Users\username\Documents\path\to\your\gcamfolder\examples\example3.RData` if you are using a Windows distribution.
+
+-   In case you are using the `gcamreport` package following the [Docker installation](#with-docker):
+
+    a)  make sure that your dataset is inside the `gcamreport` folder.
+
+    b)  make sure that you typed correctly the path to your `gcamreport` folder when generating the docker image (5th step in the [Docker section](#with-docker))
+
+    c)  make sure that you are pointing correctly to your dataset. For example, if in the `gcamreport` folder you have a folder called `examples` with your dataset `amazingDataset.RData`, you should refer to it as
+
+``` r
+# option 1: full path
+generate_report(data_path = "/app/examples/amazingDataset.RData")
+
+# option 2: partial path
+generate_report(data_path = "examples/amazingDataset.RData")
+```
+<span style="color:white">newline</span>
+
+
+
+
+#### B) Wired message when launching the UI using the Docker installation.
 
 After using the functions `generate_report()` or `launch_gcamreport_ui()` to launch the UI, you might get this message:
 
@@ -82,7 +150,7 @@ or open this url <http://localhost:4000> in your favourite browser.
 
 
 
[…]

[…]
