+++ vignettes/Dataset_Generation_Tutorial.Rmd
@@ -23,13 +23,15 @@ To see all possibilities of the `generate_report()` function, look [here](https:
 
 Assume you have a database called `myDb` in a folder named `dbFolder` and are willing to generate a standardized dataset for several scenarios: `scen1`, `scen2`, and `scen3`. With the `generate_report` function, it will be generated and automatically saved in the same folder where `myDb` is.
 
-1) Load `gcamreport` library.
+1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R Studio or Docker.
+
+2) Load `gcamreport` library.
 
 ``` r
-library(gcamreport)
+devtools::load_all()
 ```
 
-2) Store the database path and name, the query path, the desired project name, and the considered scenarios in variables. In case you do not specify the scenarios, all the scenarios in the database will be considered for reporting.
+3) Store the database path and name, the query path, the desired project name, and the desired scenarios and reporting variables in variables. In case you do not specify the scenarios, all the scenarios in the database will be considered for reporting; and if the variables are not specified, all the available reporting variables will be used. For more details about scenarios and variables specification, look at the [regions' tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-4-specify-the-regions-or-regions-groups) or at the [variables' tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-5-specify-the-variables).
 
 ``` r
 dbpath <- "/path/to/database"
@@ -42,7 +44,7 @@ Notice that the extension is included in the project name. Accepted extensions: 
 
 **Note**: in case you followed the Docker installation, you should store your database inside the `gcamreport` folder and your root directory is called `/app`. Thus, your database path will be something like `/app/path/to/database`.
 
-3) Generate the dataset until the desired year. In this example, 2050. By default is 2100 and it should be at least 2025.
+4) Generate the dataset until the desired year. In this example, 2050. By default is 2100 and it should be at least 2025.
 
 ``` r
 generate_report(db_path = dbpath, query_path = querypath, db_name = dbname, 
@@ -53,9 +55,8 @@ generate_report(db_path = dbpath, query_path = querypath, db_name = dbname,
 **Note**: project generation requires approximately 30 minutes.
 
 Notice that the dataset will automatically be saved in `.RData`, `.csv` and `.xlsx` at `/path/to/database/awesomeProj_standardized.RData`, `/path/to/database/awesomeProj_standardized.csv`, and `/path/to/database/awesomeProj_standardized.xlsx`.
-TODO specify reg, continent, variables
 
-This procedure will generate a project file at `/path/to/database/dbname_prjname.dat` with all the necessary queries that you can use as indicated in [Example 2](#example2).
+This procedure will also generate a project file at `/path/to/database/dbname_prjname.dat` with all the necessary queries that you need for reporting. You can directly use it as indicated in [Example 2](#example2).
 
 The terminal will output the performed verifications and their final status.
 
@@ -75,14 +76,17 @@ The terminal will output the performed verifications and their final status.
 Assume you have a project called `myProj.dat` and are willing to generate a standardized dataset out of it. With the `generate_report` function, it will be generated and automatically saved in the same folder where `myProj.dat` is. Notice that `myProj.dat` should have all the required queries for the standardized dataset generation. If you are not sure you have all of them or you need to generate the project, see [Example1](#example1).
 
 
-1) Load `gcamreport` library.
+1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R Studio or Docker.
+
+2) Load `gcamreport` library.
 
 ``` r
-library(gcamreport)
+devtools::load_all()
 ```
 
 
[…]
