+++ vignettes/Interactive_UI_Tutorial.Rmd
@@ -41,7 +41,7 @@ After generating an `.RData` dataset with the `generate_report()` function, simp
 # lanuch the user interface specifying the dataset variable name
 launch_gcamreport_ui(data = "report") # report is the name of the standardized 
                                       # dataset generated by generate_report() 
-                                      # function. # So, if you see it loaded in 
+                                      # function. So, if you see it loaded in 
                                       # your environment, you can easily use 
                                       # it here!! :)
 
