+++ pkgdown.yml
@@ -4,7 +4,6 @@ pkgdown_sha: ~
 articles:
   Dataset_Generation_Tutorial: Dataset_Generation_Tutorial.html
   Dataset_Generation_Tutorial_v6-0: Dataset_Generation_Tutorial_v6-0.html
-  gcamreport_v6-0: gcamreport_v6-0.html
   Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
   Interactive_UI_Tutorial_v6-0: Interactive_UI_Tutorial_v6-0.html
   Modify_Mapping_Template_Tutorial: Modify_Mapping_Template_Tutorial.html
@@ -12,7 +11,8 @@ articles:
   Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
   Step_By_Step_Full_Example_v6-0: Step_By_Step_Full_Example_v6-0.html
   Version_Guide: Version_Guide.html
-last_built: 2024-02-23T13:51Z
+  gcamreport_v6-0: gcamreport_v6-0.html
+last_built: 2024-02-23T14:18Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
