### 2024-02-26 bc3LC/gcamreport#6 @klau506
Dear @ibarraespinosa,

Thank you for your comments and performed validations. Find attached some details:

1. Regarding the First example: it might be that either you did not clone the repository, you are not using the gcamreport project, or you do not have placed the example database in the correct folder. It might also be that, when extracting the database from the zip folder, an intermediate folder appeared and the database is located into examples/database_basexdb_ref/database_basexdb_ref instead of examples/database_basexdb_ref. I have written a few lines in the tutorials to make sure no one experiences this error.
2. I have checked, and shiny is already present in the DESCRIPTION file.
3. Regarding the Third example: it might be that either you did not clone the repository or you are not using the gcamreport project. I have written a few lines in the tutorials to make sure no one experiences this error.

In order to launch the shiny app, the package relies on some files (e.g., css files) that were difficult to include as package data. Thus, the repository clone is necessary to experience the full capabilities of the tool.

Thank you for your time and hope that the clarifications are okay,

Best wishes,
Clàudia
