### 2024-03-10 bc3LC/gcamreport#6 @ibarraespinosa
Hi, I was able to download the input data and generate the gui.

I have a very basic question. Brazil should be one available region, however, it does not appear in the report

```

## -- load gcamreport library.
devtools::load_all()

## -- store the database path, name, and scenarios in a variable.
dbpath <- "examples"
dbname <- "database_basexdb_ref"
scen <- "Reference"

## -- choose a project name
prjname <- "example1.dat"

## -- generate the reporting dataset until 2050 for EU-12 and EU-15 for all the
## -- Agricultural variables, save the output in .RData, .csv and .xlsx format,
## -- and lunch the user interface

available_regions()
available_continents()

generate_report(db_path = dbpath, db_name = dbname, scenarios = scen,
                prj_name = prjname, final_year = 2050,
                desired_regions = c('Brazil', 'EU-15'),
                desired_variables = c('Agricultural*'),
                save_output = TRUE, launch_ui = TRUE)

```

![image](https://github.com/bc3LC/gcamreport/assets/27447280/68bb65c0-ca0a-41f1-8a8d-d9c008355aaf)

![image](https://github.com/bc3LC/gcamreport/assets/27447280/bfbdf55c-acd1-4f28-88ce-42f6e355ec79)
