### 2024-03-11 e7867b6
Fix cache cleaning and add UI tests
