+++ R/main.R
@@ -79,7 +79,10 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 #' @importFrom dplyr intersect setdiff if_else
 #' @export
 load_project <- function(project_path, desired_regions = "All", scenarios = NULL) {
-  print("Loading project...")
+  # rm variable "prj" from the environment if exists
+  if (exists('prj')) rm(prj, envir = .GlobalEnv)
+
+  rlang::inform("Loading project...")
 
   # load the project
   prj <- loadProject(project_path)
@@ -134,168 +137,178 @@ load_project <- function(project_path, desired_regions = "All", scenarios = NULL
 #' @import rgcam
 #' @import dplyr
 #' @export
-create_project <- function(db_path, db_name, prj_name, scenarios,
+create_project <- function(db_path, db_name, prj_name, scenarios = NULL,
                            desired_regions = "All", desired_variables = "All") {
   Internal_variable <- Variable <- required <- available_scenarios <- name <- NULL
 
-  # check if the project already exists
-  file_name <- file.path(db_path, paste(db_name, prj_name, sep = "_"))
-  if (file.exists(file_name)) {
-    load_project(file_name, desired_regions, scenarios)
+  # rm variable "prj" from the environment if exists
+  if (exists('prj')) rm(prj, envir = .GlobalEnv)
+
+  rlang::inform("Project does not exists. Creating project...")
+
+  # create the project
+  conn <- localDBConn(db_path,
+    db_name,
+    migabble = FALSE
+  )
+
+  available_scenarios <- listScenariosInDB(conn) %>%
+    pull(name)
+  # check user input
+  if (is.null(scenarios)) {
+    scenarios <- available_scenarios
   } else {
-    print("Creating project...")
-
-    # create the project
-    conn <- localDBConn(db_path,
-      db_name,
-      migabble = FALSE
-    )
-
-    available_scenarios <- listScenariosInDB(conn) %>%
-      pull(name)
-    # check user input
-    if (is.null(scenarios)) {
-      scenarios <- available_scenarios
-    } else {
-      if (length(scenarios) > length(available_scenarios)) {
-        check_scen <- setdiff(scenarios, available_scenarios)
-        tmp <- paste(check_scen, collapse = ", ")
-        if (length(check_scen) > 1) stop("The desired scenarios ", tmp, " are not present in the database.\n")
-        if (length(check_scen) == 1) stop("The desired scenario ", tmp, " is not present in the database.\n")
-      }
+    if (length(scenarios) > length(available_scenarios)) {
+      check_scen <- setdiff(scenarios, available_scenarios)
+      tmp <- paste(check_scen, collapse = ", ")
+      if (length(check_scen) > 1) stop("The desired scenarios ", tmp, " are not present in the database.\n")
+      if (length(check_scen) == 1) stop("The desired scenario ", tmp, " is not present in the database.\n")
     }
-
-    # read the queries file
-    queryFile <- "inst/extdata/queries/queries_gcamreport_gcam7.0_complete.xml"
-    queries_short <- parse_batch_query(queryFile)
-    queryFile <- "inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml"
-    queries_large <- parse_batch_query(queryFile)
-
-    # subset the queries necessary for the selected variables
-    if (!(length(desired_variables) == 1 && desired_variables == "All")) {
-      # create a mapping with the Variables, Internal variables, functions to load
-      # them, and the dependencies
-      required_internal_variables <- gcamreport::var_fun_map %>%
-        rename("Internal_variable" = "name") %>%
-        left_join(
-          gcamreport::template %>%
[…]
