+++ tests/testthat/testUI.R
@@ -26,7 +26,7 @@ test_that("Test1. test tree functions", {
   testthat::expect_equal(testResult1, testExpect1)
 
   # do_unmount_tree with regions
-  testResult1.2 <- do_unmount_tree(get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_unmount_tree_regions.RData"))), 'regions')
+  testResult1.2 <- do_unmount_tree(get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_unmount_tree_regions.RData"))), "regions")
   testExpect1.2 <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_1.2.RData")))
   testthat::expect_equal(testResult1.2, testExpect1.2)
 
@@ -36,7 +36,7 @@ test_that("Test1. test tree functions", {
   testthat::expect_equal(testResult2, testExpect2)
 
   # do_unmount_tree with variables
-  testResult2.2 <- do_unmount_tree(get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_unmount_tree_variables.RData"))), 'variables')
+  testResult2.2 <- do_unmount_tree(get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_unmount_tree_variables.RData"))), "variables")
   testExpect2.2 <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_2.2.RData")))
   testthat::expect_equal(testResult2.2, testExpect2.2)
 
@@ -85,18 +85,23 @@ test_that("Test2. error messages", {
     "Specify either the dataset or the dataset path to be considered, not both"
   )
 
-  errorMessage1 <- check_user_choices_plot(vars = 'var1', scen = c("scen1", "scen2"),
-                                           years = NULL, reg = 'Africa', grouped = TRUE)
-  testthat::expect_equal(errorMessage1, 'ERROR: Select at least one year please.')
+  errorMessage1 <- check_user_choices_plot(
+    vars = "var1", scen = c("scen1", "scen2"),
+    years = NULL, reg = "Africa", grouped = TRUE
+  )
+  testthat::expect_equal(errorMessage1, "ERROR: Select at least one year please.")
 
-  errorMessage2 <- check_user_choices_plot(vars = 'var1', scen = NULL,
-                                           years = 'year', reg = 'Africa', grouped = TRUE)
-  testthat::expect_equal(errorMessage2, 'ERROR: Select at least one scenario please.')
+  errorMessage2 <- check_user_choices_plot(
+    vars = "var1", scen = NULL,
+    years = "year", reg = "Africa", grouped = TRUE
+  )
+  testthat::expect_equal(errorMessage2, "ERROR: Select at least one scenario please.")
 
-  errorMessage3 <- check_user_choices_plot(vars = NULL, scen = 'scen1',
-                                           years = 'year', reg = 'Africa', grouped = TRUE)
-  testthat::expect_equal(errorMessage3, 'ERROR: Select at least one variable please.')
-
+  errorMessage3 <- check_user_choices_plot(
+    vars = NULL, scen = "scen1",
+    years = "year", reg = "Africa", grouped = TRUE
+  )
+  testthat::expect_equal(errorMessage3, "ERROR: Select at least one variable please.")
 })
 
 test_that("Test3. reset", {
@@ -109,13 +114,18 @@ test_that("Test3. reset", {
[…]
