### 2024-03-11 bc3LC/gcamreport#6 @klau506
Dear @ibarraespinosa,

Thank you so much for this check. I reproduced the error and found that the cache was not being properly cleaned. I amended it and now works fine, regardless of your previous gcamreport runs.

Best,
Clàudia
