### 2024-03-22 bc3LC/gcamreport#6 @klau506
Dear @ibarraespinosa ,

Thank you for your time and effort. Your thoughtful feedback has provided valuable insights that will undoubtedly enhance the quality of the work. After addressing all comments I would like to know if you have more suggestions or aspects that should be improved or discussed.

Keep in touch,
Best wishes,
Clàudia
