### 2024-03-25 bc3LC/gcamreport#6 @ibarraespinosa
I will check tonight

On Fri, Mar 22, 2024 at 4:00 AM Clàudia Rodés Bachs <
***@***.***> wrote:

> Dear @ibarraespinosa <https://github.com/ibarraespinosa> ,
>
> Thank you for your time and effort. Your thoughtful feedback has provided
> valuable insights that will undoubtedly enhance the quality of the work.
> After addressing all comments I would like to know if you have more
> suggestions or aspects that should be improved or discussed.
>
> Keep in touch,
> Best wishes,
> Clàudia
>
