### 2024-03-26 bc3LC/gcamreport#6 @ibarraespinosa
Hola,

El paquete funciona muy bien y has hecho un gran trabajo. El software lo dejare por aprobado pero tengo un comentario general. Veo en DESCRIPTION que importas un gran numero de [librerias](https://github.com/bc3LC/gcamreport/blob/gcam-v7.0/DESCRIPTION#L18). Tambien veo que importas [data.table](https://github.com/bc3LC/gcamreport/blob/gcam-v7.0/R/app_functions.R#L263) solo para retornar un objeto de clase data.table sin usar sus capacidades en operaciones.
En pocas palabras, ve la forma de reducir los paquetes.

Tambien veo que en documentation escrves importfrom data.table y luego en la funcion, as.data.table, sin data.table::as.data.table. Recuerdo que la segunda forma es la apropiada. Puedes comentar por favor?
