### 2024-03-26 bc3LC/gcamreport#6 @ibarraespinosa
DESCRIPTION mentiones LICENCE Mit + File but LICENSE file says BSD 2-Clause License. Is that equivalent with MIT?
