### 2024-03-26 bc3LC/gcamreport#6 @ibarraespinosa
As I recall, you need to have a directory named paper in your repo with the paper in it, probably built with `rticles`.
