### 2024-03-26 bc3LC/gcamreport#6 @ibarraespinosa
-[Paper](https://github.com/openjournals/joss-papers/blob/joss.05975/joss.05975/10.21105.joss.05975.pdf):

- First phrase is too long (L9-12).
- Code in summary is not needed. To that end, use the code Functionality  section, where it can be a subsection or so.
- What is the reference for gcamextractor?
- Figure 2 does not provide much information. A flow diagram would be more helpful for new users. (suggestion)
