### 2024-03-26 bc3LC/gcamreport#6 @ibarraespinosa
![image](https://github.com/bc3LC/gcamreport/assets/27447280/6ff0fbaa-38c5-4334-a0fc-865e960d835f)

![image](https://github.com/bc3LC/gcamreport/assets/27447280/c37aa695-a966-4388-92d8-2d98896a2b7a)

I could not access the help index, is that OK?
