### 2024-03-26 21a1643
Review: simplify necessary pkgs
