+++ NAMESPACE
@@ -98,7 +98,6 @@ export(update_user_choices_plot)
 import(dplyr)
 import(ggplot2)
 import(rgcam)
-importFrom(data.table,as.data.table)
 importFrom(dplyr,anti_join)
 importFrom(dplyr,bind_rows)
 importFrom(dplyr,distinct)
