+++ Dockerfile
@@ -20,7 +20,6 @@ RUN R -e "remotes::install_github('JGCRI/rgcam')"
 RUN R -e "install.packages('knitr')"
 RUN R -e "install.packages('markdown')"
 RUN R -e "install.packages('readr')"
-RUN R -e "install.packages('stringi')"
 RUN R -e "install.packages('rlang')"
 RUN R -e "install.packages('readr')"
 RUN R -e "install.packages('data.table')"
