+++ paper/paper.md
@@ -1,0 +1,107 @@
+---
+title: 'gcamreport: An R tool to process and standardize GCAM outputs'
+tags:
+- R
+- analysis
+- IAMC
+- GCAM
+date: "24 July 2023"
+affiliations:
+- name: Basque Centre for Climate Change (BC3), Leioa, Spain
+  index: 1
+- name: Center for Global Sustainability, School of Public Policy, University of Maryland, College Park, MD, USA
+  index: 2
+- name: Pacific Northwest National Laboratory's Joint Global Change Research Institute, College Park, MD, USA
+  index: 3
+authors:
+- name: "Cl&#224;udia Rod&#233;s-Bachs"
+  orcid: "0000-0001-6696-7685"
+  affiliation: 1
+- name: Jon Sampedro
+  orcid: "0000-0002-2277-1530"
+  affiliation: 1
+- name: "Dirk-Jan Van de Ven"
+  orcid: "0000-0001-9120-564X"
+  affiliation: 1
+- name: "Ryna Yiyun Cui"
+  orcid: "0000-0002-1186-8230"
+  affiliation: 2
+- name: "Alicia Zhao"
+  orcid: "0000-0002-6054-8671"
+  affiliation: 2
+- name: "Matthew Zwerling"
+  affiliation: 2
+- name: "Zarrar Khan"
+  orcid: "0000-0002-8147-8553"
+  affiliation: 3
+bibliography: paper.bib
+---
+
+# Summary
+
+There is an urgent need to perform multi-model studies to characterize uncertainty arising from models' heterogeneity with the aim of building a more reliable and transparent framework to inform policy makers in the design and implementation of climate policies [@guivarch_using_2022]. In response to this challenge, multiple institutes and organizations have adopted the standardized data template developed by the Integrated Assessment Modeling Consortium (IAMC). This template is maintained by the International Institute for Applied Systems Analysis (IIASA) and aims to standardize and facilitate model intercomparison exercises. For the latest Assessment Report (AR6), the Intergovernmental Panel on Climate Change (IPCC) required all contributors to homogenize their data to enable comparisons and ensure full transparency [@krey_annex_2014]. This practice has set the foundation for a new open management of the outputs in the area of global scenario analysis.  
+
+In the case of the Global Change Analysis Model (GCAM) [@calvin2019], a well-regarded model that has been extensively used for different international and national scenario analysis, the harmonization code has never been documented nor standardized, making it difficult to reproduce outputs and hindering the transparency of results. To overcome these limitations, we have developed `gcamreport`, an R package that systematizes the transformations of GCAM outputs, generates figures to facilitate the analysis of the results, and allows user interaction with the produced outputs. Furthermore, the tool can be used embedded in a Docker image, which allows users to use the package in a virtual environment without having to install any specific software or library. Finally, each `gcamreport` release is linked to either a version of GCAM or a study in which GCAM was used, ensuring reproducibility, interoperability, accessibility, and findability, which is in line with the well-known open science principles FAIR and TRUST [@wilkinson2016; @lin2020].
+
+
+![**Figure 1:** Structure of the `gcamreport` package.\label{fig:scheme}](scheme_fig4_crop.png){width="75%"}
+
+The `gcamreport` package is accessible online at the public domain <https://github.com/bc3LC/gcamreport>. To run it to generate and save the dataset meeting the IAMC requirements and to launch the user interface, you can follow a detailed [step-by-step tutorial](https://bc3lc.github.io/gcamreport/articles/Step_By_Step_Full_Example.html) or this simplified code once the package has been cloned:
+
+``` r
+# 1) open the gcamreport project
+
[…]
