+++ DESCRIPTION
@@ -16,6 +16,7 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.1
 Imports:
+    usethis,
   	markdown,
     rlang,
     readr,
@@ -38,7 +39,6 @@ Imports:
     rrapply (>= 1.2.6),
     rpackageutils
 Suggests:
-    usethis,
     knitr,
     rprojroot,
     rmarkdown,
