+++ paper/paper.md
@@ -6,13 +6,10 @@ tags:
 - IAMC
 - GCAM
 date: "24 July 2023"
-affiliations:
-- name: Basque Centre for Climate Change (BC3), Leioa, Spain
-  index: 1
-- name: Center for Global Sustainability, School of Public Policy, University of Maryland, College Park, MD, USA
-  index: 2
-- name: Pacific Northwest National Laboratory's Joint Global Change Research Institute, College Park, MD, USA
-  index: 3
+output:
+  pdf_document: default
+  html_document:
+    df_print: paged
 authors:
 - name: "Cl&#224;udia Rod&#233;s-Bachs"
   orcid: "0000-0001-6696-7685"
@@ -23,28 +20,52 @@ authors:
 - name: "Dirk-Jan Van de Ven"
   orcid: "0000-0001-9120-564X"
   affiliation: 1
-- name: "Ryna Yiyun Cui"
+- name: Ryna Yiyun Cui
   orcid: "0000-0002-1186-8230"
   affiliation: 2
-- name: "Alicia Zhao"
+- name: Alicia Zhao
   orcid: "0000-0002-6054-8671"
   affiliation: 2
-- name: "Matthew Zwerling"
+- name: Matthew Zwerling
   affiliation: 2
-- name: "Zarrar Khan"
+- name: Zarrar Khan
   orcid: "0000-0002-8147-8553"
   affiliation: 3
 bibliography: paper.bib
+affiliations:
+- name: Basque Centre for Climate Change (BC3), Leioa, Spain
+  index: 1
+- name: Center for Global Sustainability, School of Public Policy, University of Maryland,
+    College Park, MD, USA
+  index: 2
+- name: Pacific Northwest National Laboratory's Joint Global Change Research Institute,
+    College Park, MD, USA
+  index: 3
 ---
 
 # Summary
 
-There is an urgent need to perform multi-model studies to characterize uncertainty arising from models' heterogeneity with the aim of building a more reliable and transparent framework to inform policy makers in the design and implementation of climate policies [@guivarch_using_2022]. In response to this challenge, multiple institutes and organizations have adopted the standardized data template developed by the Integrated Assessment Modeling Consortium (IAMC). This template is maintained by the International Institute for Applied Systems Analysis (IIASA) and aims to standardize and facilitate model intercomparison exercises. For the latest Assessment Report (AR6), the Intergovernmental Panel on Climate Change (IPCC) required all contributors to homogenize their data to enable comparisons and ensure full transparency [@krey_annex_2014]. This practice has set the foundation for a new open management of the outputs in the area of global scenario analysis.  
+There is an urgent need for multi-model studies to characterize uncertainty arising from model heterogeneity. These studies aim to build a more reliable and transparent framework, informing policymakers in the design and implementation of climate policies [@guivarch_using_2022]. In response to this challenge, multiple institutes and organizations have adopted the standardized data template developed by the Integrated Assessment Modeling Consortium (IAMC). This template is maintained by the International Institute for Applied Systems Analysis (IIASA) and aims to standardize and facilitate model intercomparison exercises. For the latest Assessment Report (AR6), the Intergovernmental Panel on Climate Change (IPCC) required all contributors to homogenize their data to enable comparisons and ensure full transparency [@krey_annex_2014]. This practice has set the foundation for a new open management of the outputs in the area of global scenario analysis.  
 
[…]
