+++ paper/paper.bib
@@ -1,3 +1,12 @@
+@software{waite_gcamextractor_2023,
+  author       = {Taryn Waite and Leeya Pressburger and Mengqi Zhao and Zarrar Khan},
+  title        = {{JGCRI/gcamextractor: v1.1.0}},
+  year         = 2023,
+  version      = {v1.1.0},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.8132290},
+  url          = {https://doi.org/10.5281/zenodo.8132290}
+}
 
 @misc{krey_iamc_2023,
 	address = {Online},
