### 2024-03-26 bc3LC/gcamreport#6 @klau506
Dear @ibarraespinosa ,

Thank you for your checks! Some answers below:
> Hola,
>
> El paquete funciona muy bien y has hecho un gran trabajo. El software lo dejare por aprobado pero tengo un comentario general. Veo en DESCRIPTION que importas un gran numero de [librerias](https://github.com/bc3LC/gcamreport/blob/gcam-v7.0/DESCRIPTION#L18). Tambien veo que importas [data.table](https://github.com/bc3LC/gcamreport/blob/gcam-v7.0/R/app_functions.R#L263) solo para retornar un objeto de clase data.table sin usar sus capacidades en operaciones. En pocas palabras, ve la forma de reducir los paquetes.

Good point! Thank you! I have erased some packages that could be easily substituted by others, and have moved some that were only used for testing purposes to the _suggested packages_ section

> Tambien veo que en documentation escrves importfrom data.table y luego en la funcion, as.data.table, sin data.table::as.data.table. Recuerdo que la segunda forma es la apropiada. Puedes comentar por favor?

The code was initially written with this notation, but another reviewer, bpbond, pointed out [in this issue](https://github.com/bc3LC/gcamreport/issues/5#:~:text=Since%20you%27re%20using%20Roxygen%2C%20use%20importFrom%20in%20function%20headers%20to%20make%20functions%20available%20(and%20eliminate%20all%20the%20dplyr%3A%3A%2Dtype%20namespace%20qualifiers%2C%20too%2C%20improving%20readability)) that `import from` was best practice. That's why I have been following this notation/format accross the package

> DESCRIPTION mentiones LICENCE Mit + File but LICENSE file says BSD 2-Clause License. Is that equivalent with MIT?

You are right, the LICENCE didn't match. I re-uploaded the Mit+ LICENCE

> As I recall, you need to have a directory named paper in your repo with the paper in it, probably built with `rticles`.

Sure! The paper directory was in [another branch](https://github.com/bc3LC/gcamreport/tree/paper), but I have [moved ](https://github.com/bc3LC/gcamreport/tree/paper/paper)it to the main branch as suggested. I will be easier to find ;)

> -[Paper](https://github.com/openjournals/joss-papers/blob/joss.05975/joss.05975/10.21105.joss.05975.pdf):
>
> * First phrase is too long (L9-12).
> * Code in summary is not needed. To that end, use the code Functionality  section, where it can be a subsection or so.
> * What is the reference for gcamextractor?
> * Figure 2 does not provide much information. A flow diagram would be more helpful for new users. (suggestion)

Thank you for the manuscript suggestions.
- The first sentence was re-written
- The code example was moved to the Functionality section, where two subsections were created
- The `gcamextractor` was properly cited
- Figure 2 was modified to show the user interface screenshot and the user management workflow. Thank you for the suggestion, it will help to understand better the usage workflow

> I could not access the help index, is that OK?

I'm sorry, but where did you see this?

Thank you so much,
Best,
Clàudia
