+++ R/main.R
@@ -9,18 +9,26 @@
 #' @param desired_regions desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
 #' To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
 #' In case the project dataset needs to be created, it will be produced with only the specified regions.
+#' @param queries_nonCO2_file full path to an xml query file (including file name and extension) designed to load long nonCO2 queries:
+#' "nonCO2 emissions by sector (excluding resource production)" and "nonCO2 emissions by region". By default it points to the
+#' gcamreport nonCO2 queries file, compatible with the latest GCAM version and necessary to report some of the standardized variables.
 #' @return dataframe with the data from the query.
 #' @importFrom rgcam addSingleQuery localDBConn
 #' @importFrom xml2 read_xml xml_find_first
 #' @importFrom dplyr bind_rows
 #' @export
-data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regions = "All") {
+data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regions = "All",
+                       queries_nonCO2_file = gcamreport::queries_nonCO2) {
   if (identical(desired_regions, "All")) {
     desired_regions <- NULL
   }
 
   dt <- data.frame()
-  xml <- read_xml("inst/extdata/queries/queries_gcamreport_gcam7.0_nonCO2.xml")
+  if (is.list(queries_nonCO2_file)) {
+    xml <- transform_to_xml(queries_nonCO2_file)
+  } else {
+    xml <- read_xml(queries_nonCO2_file)
+  }
   qq <- xml_find_first(xml, paste0("//*[@title='", type, "']"))
 
   for (sc in scenarios) {
@@ -130,12 +138,19 @@ load_project <- function(project_path, desired_regions = "All", scenarios = NULL
 #' In case the project dataset needs to be created, it will be produced with only the specified variables. ATTENTION:
 #' the global variables such as "Emissions" will be computed considering only the selected variables, for instance "Emissions|CO2",
 #' and will no account for other variables, such as "Emissions|CH4" or "Emissions|NH3".
-#' @return loaded project into global environment.
+#' @param queries_general_file full path to a general xml query file (including file name and extension). By default it points to the
+#' gcamreport general queries file, compatible with the latest GCAM version and able to report all standardized variables.
+#' @param queries_nonCO2_file full path to an xml query file (including file name and extension) designed to load long nonCO2 queries:
+#' "nonCO2 emissions by sector (excluding resource production)" and "nonCO2 emissions by region". By default it points to the
+#' gcamreport nonCO2 queries file, compatible with the latest GCAM version and necessary to report some of the standardized variables.
+#' @return loaded project into global environment and local saved project.
 #' @import rgcam
 #' @import dplyr
 #' @export
-create_project <- function(db_path, db_name, prj_name, scenarios,
-                           desired_regions = "All", desired_variables = "All") {
+create_project <- function(db_path, db_name, prj_name, scenarios = NULL,
+                           desired_regions = "All", desired_variables = "All",
+                           queries_general_file = gcamreport::queries_general,
+                           queries_nonCO2_file = gcamreport::queries_nonCO2) {
   Internal_variable <- Variable <- required <- available_scenarios <- name <- NULL
 
   # check if the project already exists
@@ -166,10 +181,16 @@ create_project <- function(db_path, db_name, prj_name, scenarios,
     }
 
     # read the queries file
-    queryFile <- "inst/extdata/queries/queries_gcamreport_gcam7.0_complete.xml"
-    queries_short <- parse_batch_query(queryFile)
[…]
