+++ man/generate_report.Rd
@@ -16,7 +16,9 @@ generate_report(
   desired_continents = "All",
   save_output = TRUE,
   output_file = NULL,
-  launch_ui = TRUE
+  launch_ui = TRUE,
+  queries_general_file = gcamreport::queries_general,
+  queries_nonCO2_file = gcamreport::queries_nonCO2
 )
 }
 \arguments{
@@ -56,9 +58,17 @@ database or the project file, and using a default name containing the project na
 In case of specifying the `output_file`, introduce a whole path (e.g. /path/to/output/fileName) without extension tag, it will be automatically added.}
 
 \item{launch_ui}{if TRUE, launch User Interface, Do not launch it otherwise.}
+
+\item{queries_general_file}{full path to a general xml query file (including file name and extension). By default it points to the
+gcamreport general queries file, compatible with the latest GCAM version and able to report all standardized variables.}
+
+\item{queries_nonCO2_file}{full path to an xml query file (including file name and extension) designed to load long nonCO2 queries:
+"nonCO2 emissions by sector (excluding resource production)" and "nonCO2 emissions by region". By default it points to the
+gcamreport nonCO2 queries file, compatible with the latest GCAM version and necessary to report some of the standardized variables.}
 }
 \value{
-RData, CSV, and XLSX datafile with the desired variables & launches user interface.
+RData, CSV, and XLSX saved datafiles with the desired standardized variables, launches user interface, and save the rgcam
+project file (if created).
 }
 \description{
 Main function. Creates/loads a GCAM project, standardizes the data and saves it
