+++ man/transform_to_xml.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/functions.R
+\name{transform_to_xml}
+\alias{transform_to_xml}
+\title{transform_to_xml}
+\usage{
+transform_to_xml(parsed_queries_list)
+}
+\arguments{
+\item{parsed_queries_list}{parsed list of queries}
+}
+\value{
+xml document containing the given queries list
+}
+\description{
+Return xml document from a given queries list
+}
+\keyword{internal}
+\keyword{utils}
