+++ man/queries_general.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{queries_general}
+\alias{queries_general}
+\title{queries_general}
+\format{
+vector
+}
+\source{
+local
+}
+\usage{
+queries_general
+}
+\description{
+gcamreport queries compatible with GCAM 7.0 version.
+Contain all queries except for nonCO2 queries
+}
+\examples{
+\dontrun{
+library(gcamreport)
+gcamreport::queries_general
+}
+}
+\keyword{datasets}
