+++ NAMESPACE
@@ -94,6 +94,7 @@ export(load_query)
 export(load_variable)
 export(reset_first_load)
 export(start_with_pattern)
+export(transform_to_xml)
 export(update_user_choices_plot)
 import(dplyr)
 import(ggplot2)

[…]
