### 2024-04-04 5e41db1
Add tests for queries specification parameters
