+++ tests/testthat/testUI.R
@@ -142,8 +142,39 @@ test_that("Test5. do_data_sample", {
   sel_vars <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_5.sel_vars.RData")))
   sel_years <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_5.sel_years.RData")))
   testResult <- do_data_sample(sdata, sel_scen, sel_years, sel_cols, sel_vars, sel_reg, 0, 0)
-  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_5.data_sample.RData")))
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_5.1.RData")))
   testthat::expect_equal(testResult, testExpect)
+
+  testResult <- do_data_sample(sdata, sel_scen, sel_years, sel_cols, sel_vars, sel_reg, 1, 0)
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_5.2.RData")))
+  testthat::expect_equal(testResult, testExpect)
+
+  testResult <- do_data_sample(sdata, sel_scen, sel_years, sel_cols, sel_vars, sel_reg, 1, 1)
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_5.3.RData")))
+  testthat::expect_equal(testResult, testExpect)
+
+  testResult <- do_data_sample(sdata, sel_scen, sel_years, sel_cols, sel_vars, sel_reg, 2, 1)
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_5.4.RData")))
+  testthat::expect_equal(testResult, testExpect)
+
+  testResult <- do_data_sample(sdata, sel_scen, sel_years, sel_cols, sel_vars, sel_reg, 1, 2)
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_5.5.RData")))
+  testthat::expect_equal(testResult, testExpect)
+})
+
+
+test_that("Test6. update_user_choices_plot", {
+
+  # update_user_choices_plot
+  selected_scen <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_6.selected_scen.RData")))
+  selected_years <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_6.selected_years.RData")))
+  sidebarItemExpanded <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_6.sidebarItemExpanded.RData")))
+  tree_regions <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_6.tree_regions.RData")))
+  tree_variables <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test_ui_6.tree_variables.RData")))
+  testResult <- update_user_choices_plot(selected_scen, selected_years, tree_regions, tree_variables, sidebarItemExpanded)
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/test_ui_6.1.RData")))
+  testthat::expect_equal(testResult, testExpect)
+
 })
 
 
