+++ tests/testthat/testReporter_v7.0.R
@@ -589,3 +589,26 @@ test_that("Test12_v7. other functions", {
   testthat::expect_equal(testResult, testExpect)
 
 })
+
+test_that("Test13_v7. specify queries", {
+
+  # transform_to_xml ancillary function
+  testResult <- transform_to_xml(gcamreport::queries_nonCO2)
+  testExpect <- xml2::read_xml(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test13.1.xml"))
+  testthat::expect_equal(testResult, testExpect)
+
+  # generate standardize report specifying the queries file
+  db_path <- file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0")
+  db_name <- "database_basexdb_ref"
+  prj_name <- "gcamv7.0_test_specify_queries.dat"
+  scenarios <- "Reference"
+  generate_report(db_path = db_path, db_name = db_name, prj_name = prj_name,
+                  scenarios = scenarios, final_year = 2050, desired_variables = c('Price|Carbon*'),
+                  save_output = T, launch_ui = F,
+                  queries_general_file = file.path(rprojroot::find_root(rprojroot::is_testthat), "inst/extdata/queries/queries_gcamreport_gcam7.0_complete.xml"))
+  assign('testExpect',
+         rgcam::loadProject(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test13.2_prj.dat")))
+  testthat::expect_equal(prj$Reference, testExpect$Reference)
+
+})
+
