+++ .github/workflows/test_coverage.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - gcam-v7.0
+      - dev
 
 jobs:
   test:
