+++ .github/workflows/docker_impl.yaml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - gcam-v7.0
+      - dev
 
 jobs:
   build:
