### 2024-04-04 761f6d4
Add tests for queries specification parameters
