+++ tests/testthat/testReporter_v7.0.R
@@ -474,3 +474,76 @@ test_that("Test11_v7. scenarios", {
   testResult <- rgcam::listScenarios(prj)
   testthat::expect_equal("CP_EI_recovery", testResult)
 })
+
+
+test_that("Test12_v7. other functions", {
+  # gather_map
+  co2_sector_map <- read.csv(file.path(rprojroot::find_root(rprojroot::is_testthat), "inst/extdata/mappings", "CO2_sector_map.csv"),
+                             skip = 1, na = "",
+                             stringsAsFactors = FALSE
+  ) %>% gather_map()
+
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test12.1.RData")))
+  testthat::expect_equal(co2_sector_map, testExpect)
+
+  # approx_fun
+  expect_error(
+    approx_fun(2030, 3, rule = 3),
+    "Use fill_exp_decay_extrapolate!"
+  )
+
+  # get_iron_steel_map & get_co2_iron_steel
+  generate_report(
+    prj_name = file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test7.dat"),
+    final_year = 2050,
+    scenarios = "Reference",
+    desired_regions = "USA",
+    desired_variables = "Emissions|CO2|Energy|Demand|Industry|Steel",
+    launch_ui = FALSE
+  )
+
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test12.2.RData")))
+  testthat::expect_equal(report, testExpect)
+
+  # get_ghg
+  generate_report(
+    prj_name = file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0/test7.dat"),
+    final_year = 2050,
+    scenarios = "Reference",
+    desired_regions = "USA",
+    desired_variables = "Emissions|Kyoto Gases*",
+    launch_ui = FALSE
+  )
+
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test12.3.RData")))
+  testthat::expect_equal(report, testExpect)
+
+  # get_regional_emission
+  testResult <- get_regional_emission()
+  testExpect <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test12.4.RData")))
+  testthat::expect_equal(testResult, testExpect)
+
+})
+
+test_that("Test13_v7. specify queries", {
+
+  # transform_to_xml ancillary function
+  testResult <- transform_to_xml(gcamreport::queries_nonCO2)
+  testExpect <- xml2::read_xml(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/v_7.0/result_test13.1.xml"))
+  testthat::expect_equal(testResult, testExpect)
+
+  # generate standardize report specifying the queries file
+  db_path <- file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs/v_7.0")
+  db_name <- "database_basexdb_ref"
+  prj_name <- "gcamv7.0_test_specify_queries.dat"
+  scenarios <- "Reference"
+  generate_report(db_path = db_path, db_name = db_name, prj_name = prj_name,
+                  scenarios = scenarios, final_year = 2050, desired_variables = c('Price|Carbon*'),
+                  save_output = T, launch_ui = F,
+                  queries_general_file = file.path(rprojroot::find_root(rprojroot::is_testthat), "inst/extdata/queries/queries_gcamreport_gcam7.0_complete.xml"))
[…]
