+++ .github/workflows/test_coverage.yml
@@ -3,7 +3,8 @@ name: test_coverage
 on:
   push:
     branches:
-      - gcam-v6.0
+      - gcam-v7.0
+      - dev
 
 jobs:
   test:
