+++ vignettes/Dataset_Generation_Tutorial.Rmd
@@ -7,7 +7,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-To see all possibilities of the `generate_report()` function, look [here](https://bc3lc.github.io/gcamreport/reference/generate_report.html).
+To see a formal description of the `generate_report()` function, look [here](https://bc3lc.github.io/gcamreport/reference/generate_report.html).
 
 <span style="color:white">newline</span>
 
@@ -21,30 +21,38 @@ To see all possibilities of the `generate_report()` function, look [here](https:
 
 ## Example 1: generate a dataset from a database {#example1}
 
-Assume you have a database called `myDb` in a folder named `dbFolder` and are willing to generate a standardized dataset for several scenarios: `scen1`, `scen2`, and `scen3`. With the `generate_report` function, it will be generated and automatically saved in the same folder where `myDb` is.
+Suppose you have a database named `myDb` in a folder named `dbFolder` and you want to generate a standardized dataset for several scenarios: `scen1`, `scen2` and `scen3` using a `new_general_queries_file`. The `generate_report` function will generate it and automatically save it in the same folder where `myDb` is located.
 
-1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R Studio or Docker.
 
-2) Load `gcamreport` library.
+1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R or Docker.
+
+2) Load the `gcamreport` library. If you are using Rstudio or Docker, run
 
 ``` r
 devtools::load_all()
 ```
 
-3) Store the database path and name, the query path, the desired project name, and the desired scenarios and reporting variables in variables. In case you do not specify the scenarios, all the scenarios in the database will be considered for reporting; and if the variables are not specified, all the available reporting variables will be used. For more details about scenarios and variables specification, look at the [regions' tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-4-specify-the-regions-or-regions-groups) or at the [variables' tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-5-specify-the-variables).
+and if you are using R, run
+
+``` r
+library(gcamreport)
+```
+
+3) Store the database path and name, the query path, the desired project name, and the desired scenarios and reporting variables in variables. In case you do not specify the scenarios, all the scenarios in the database will be considered for reporting; and if the variables are not specified, all the available reporting variables will be used. For more details about scenarios and variables specification, look at the [regions' tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-4-specify-the-regions-or-regions-groups) or at the [variables' tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-5-specify-the-variables). If you want to specify the query files, look at the [query files tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#example-6-specify-the-query-files).
 
 ``` r
 dbpath <- "/path/to/database"
 dbname <- "gcamdb_name"
 prjname <- "awesomeProj.dat"
-scen <- c('scen1', 'scen2', 'scen3')
+scen <- c("scen1", "scen2", "scen3")
 ```
 
-Notice that the extension is included in the project name. Accepted extensions: `.dat` & `.proj`.
+Notice that the extension is included in the project name. Accepted extensions are `.dat` & `.proj`.
 
-**Note**: in case you followed the Docker installation, you should store your database inside the `gcamreport` folder and your root directory is called `/app`. Thus, your database path will be something like `/app/path/to/database`.
[…]
