+++ vignettes/Step_By_Step_Full_Example.Rmd
@@ -7,7 +7,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-In this tutorial, we will 1) generate a dataset from a provided GCAM7.0 database and run the associated user interface and 2) generate a dataset from a provided project and run the corresponding user interface. To know more about the `gcamreport` package possibilities, look at these tutorials: [dataset generation tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html) and [user interface tutorial](https://bc3lc.github.io/gcamreport/articles/Interactive_UI_Tutorial.html).
+In this tutorial, we will 1) generate a dataset from a provided GCAM7.0 database and launch the associated user interface, 2) generate a dataset from a provided project and launch the corresponding user interface, and 3) launch the user interface from a provided standardized dataset. To know more about the `gcamreport` package possibilities, look at these tutorials: [dataset generation tutorial](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html) and [user interface tutorial](https://bc3lc.github.io/gcamreport/articles/Interactive_UI_Tutorial.html).
 
 <span style="color:white">newline</span>
 
@@ -21,16 +21,25 @@ In this tutorial, we will 1) generate a dataset from a provided GCAM7.0 database
 
 ## Example 1: step-by-step standardized dataset generation from a provided database
 
-1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R Studio or Docker.
+1) Follow the [installation guide](https://bc3lc.github.io/gcamreport/index.html#installation-guide) either with R or Docker.
 
 2) Download the example [GCAM7.0 database](https://zenodo.org/records/10258919), store it inside `gcamreport/examples`, and unpack it. Make sure that the database is directly located in `gcamreport/examples/database_basexdb_ref` and an intermediate folder has not appeared.
 
-3) Generate the standardized dataset:
+3) Load the `gcamreport` library. If you are using Rstudio or Docker, run
 
 ``` r
-## -- load gcamreport library.
 devtools::load_all()
+```
 
+and if you are using R, run
+
+``` r
+library(gcamreport)
+```
+
+4) Generate the standardized dataset:
+
+``` r
 ## -- store the database path, name, and scenarios in a variable.
 dbpath <- "examples"
 dbname <- "database_basexdb_ref"
@@ -49,18 +58,20 @@ generate_report(db_path = dbpath, db_name = dbname, scenarios = scen,
                 save_output = TRUE, launch_ui = TRUE)
 ```
 
-In case you expirence some trouble, check this [troubleshooting section](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#troubleshooting-for-the-generate_report-function).
+In case you experience some trouble, check this [troubleshooting section](https://bc3lc.github.io/gcamreport/articles/Dataset_Generation_Tutorial.html#troubleshooting-for-the-generate_report-function). 
 
 
-4) Once the project has been generated and the standardized dataset created, the user interface will prompt. If you are using R Studio, it will automatically open. For a better experience, click the `open in browser` button. If you are using Docker, either type <http://localhost:4000> in the browser or go to Docker Desktop and click the last started port.
+5) Once the project has been generated and the standardized dataset created, the user interface will prompt. If you are using Rstudio, it will automatically open. For a better experience, click the `open in browser` button. If you are using Docker, either type <http://localhost:4000> in the browser or go to Docker Desktop and click the last started port.
 
 <img src="https://raw.githubusercontent.com/bc3LC/gcamreport/gcam-v7.0/vignettes/readme_fig/shiny_error1.png" title="Click the last started docker port" alt="UI error" width="80%" height="80%"/>
 
[…]
