### 2024-04-05 2b3409d
Update documentation with query files selection
