+++ R/main.R
@@ -39,8 +39,8 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 
       prj_tmp <- addSingleQuery(
         conn = localDBConn(db_path,
-          db_name,
-          migabble = FALSE
+                           db_name,
+                           migabble = FALSE
         ),
         proj = prj_name,
         qn = type,
@@ -87,7 +87,10 @@ data_query <- function(type, db_path, db_name, prj_name, scenarios, desired_regi
 #' @importFrom dplyr intersect setdiff if_else
 #' @export
 load_project <- function(project_path, desired_regions = "All", scenarios = NULL) {
-  print("Loading project...")
+  # rm variable "prj" from the environment if exists
+  if (exists("prj")) rm(prj, envir = .GlobalEnv)
+
+  rlang::inform("Loading project...")
 
   # load the project
   prj <- loadProject(project_path)
@@ -153,10 +156,22 @@ create_project <- function(db_path, db_name, prj_name, scenarios = NULL,
                            queries_nonCO2_file = gcamreport::queries_nonCO2) {
   Internal_variable <- Variable <- required <- available_scenarios <- name <- NULL
 
-  # check if the project already exists
-  file_name <- file.path(db_path, paste(db_name, prj_name, sep = "_"))
-  if (file.exists(file_name)) {
-    load_project(file_name, desired_regions, scenarios)
+  # rm variable "prj" from the environment if exists
+  if (exists("prj")) rm(prj, envir = .GlobalEnv)
+
+  rlang::inform("Project does not exists. Creating project...")
+
+  # create the project
+  conn <- localDBConn(db_path,
+                      db_name,
+                      migabble = FALSE
+  )
+
+  available_scenarios <- listScenariosInDB(conn) %>%
+    pull(name)
+  # check user input
+  if (is.null(scenarios)) {
+    scenarios <- available_scenarios
   } else {
     check_scen <- setdiff(scenarios, available_scenarios)
     tmp <- paste(check_scen, collapse = ", ")
@@ -164,159 +179,153 @@ create_project <- function(db_path, db_name, prj_name, scenarios = NULL,
     if (length(check_scen) == 1) stop("The desired scenario ", tmp, " is not present in the database.\n")
   }
 
-    # create the project
-    conn <- localDBConn(db_path,
-      db_name,
-      migabble = FALSE
-    )
-
-    available_scenarios <- listScenariosInDB(conn) %>%
-      pull(name)
-    # check user input
-    if (is.null(scenarios)) {
-      scenarios <- available_scenarios
-    } else {
-      if (length(scenarios) > length(available_scenarios)) {
-        check_scen <- setdiff(scenarios, available_scenarios)
-        tmp <- paste(check_scen, collapse = ", ")
-        if (length(check_scen) > 1) stop("The desired scenarios ", tmp, " are not present in the database.\n")
-        if (length(check_scen) == 1) stop("The desired scenario ", tmp, " is not present in the database.\n")
-      }
-    }
-
-    # read the queries file
-    if (is.list(queries_general_file)) {
-      queries_short <- queries_general_file
-    } else {
-      queries_short <- parse_batch_query(queries_general_file)
-    }
-    if (is.list(queries_nonCO2_file)) {
-      queries_large <- queries_nonCO2_file
-    } else {
-      queries_large <- parse_batch_query(queries_nonCO2_file)
-    }
-
-    # subset the queries necessary for the selected variables
-    if (!(length(desired_variables) == 1 && desired_variables == "All")) {
-      # create a mapping with the Variables, Internal variables, functions to load
-      # them, and the dependencies
-      required_internal_variables <- gcamreport::var_fun_map %>%
[…]
