+++ paper/paper.bib
@@ -5,16 +5,16 @@
   version      = {v1.1.0},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.8132290},
-  url          = {https://doi.org/10.5281/zenodo.8132290}
+  url          = {https://doi.org/10.5281/zenodo.8132290},
 }
 
 @misc{krey_iamc_2023,
-	address = {Online},
-	title = {{IAMC} {Webinar}: {Standardization} of scenario data collection via the {IAMC} data template: past developments and future challenges},
-	url = {https://www.iamconsortium.org/event/iamc-webinar-standardization-of-scenario-data-collection-via-the-iamc-data-template-past-developments-and-future-challenges-volker-krey/},
-	author = {Krey, Volker},
-	month = jun,
-	year = {2023},
+    author = {Krey, Volker},
+    title = {{IAMC} {Webinar}: {Standardization} of scenario data collection via the {IAMC} data template: past developments and future challenges},
+    howpublished = {Online},
+    month = jun,
+    year = {2023},
+    url = {https://www.iamconsortium.org/event/iamc-webinar-standardization-of-scenario-data-collection-via-the-iamc-data-template-past-developments-and-future-challenges-volker-krey/},
 }
 
 @article{richters_ngfs_2022,
@@ -22,6 +22,7 @@
 	author = {Richters, O and Bertram, C and Kriegler, E and Al Khourdajie, A and Cui, R and Edmonds, J and Hackstock, P and Holland, D and Hurst, I and Kikstra, J},
 	year = {2022},
 	note = {Publisher: Zenodo},
+	doi = {10.5281/zenodo.7085661},
 }
 
 @article{masson-delmotte_climate_2021,
@@ -31,6 +32,7 @@
 	author = {Masson-Delmotte, Valérie and Zhai, Panmao and Pirani, Anna and Connors, Sarah L and Péan, Clotilde and Berger, Sophie and Caud, Nada and Chen, Y and Goldfarb, L and Gomis, MI},
 	year = {2021},
 	note = {Publisher: Cambridge University Press Cambridge, UK},
+	doi = {10.1017/9781009157896},
 }
 
 @article{masson-delmotte_ipcc_2018,
@@ -38,7 +40,7 @@
 	language = {en},
 	author = {Masson-Delmotte, Valérie and Pörtner, Hans-Otto and Skea, Jim and Zhai, Panmao and Roberts, Debra and Shukla, Priyadarshi R and Pirani, Anna and Pidcock, Roz and Chen, Yang and Lonnoy, Elisabeth and Moufouma-Okia, Wilfran and Péan, Clotilde and Connors, Sarah and Matthews, J B Robin and Zhou, Xiao and Gomis, Melissa I and Maycock, Tom and Tignor, Melinda and Waterfield, Tim},
 	year = {2018},
-	file = {Masson-Delmotte et al. - An IPCC Special Report on the impacts of global wa.pdf:C\:\\Users\\claudia.rodes\\Zotero\\storage\\W5JPBQMT\\Masson-Delmotte et al. - An IPCC Special Report on the impacts of global wa.pdf:application/pdf},
+	doi = {10.1017/9781009157940},
 }
 
 @article{nikas_perspective_2021,
@@ -54,7 +56,6 @@
 	month = jan,
 	year = {2021},
 	pages = {119153},
[…]
