### 2024-04-05 bc3LC/gcamreport#6 @ibarraespinosa
Hi, I was able to see the help index. I checked the documentation and functions, and paper, and all seems ok to me
