### 2024-04-08 #5975 @klau506
Dear @martinfleis,

I updated all the DOIs in the paper. Three references don't have a DOI, but URL is added. Thanks!
