+++ pkgdown.yml
@@ -1,18 +1,18 @@
 pandoc: 3.1.11
-pkgdown: 2.0.7
+pkgdown: 2.0.8
 pkgdown_sha: ~
 articles:
-  Dataset_Generation_Tutorial: Dataset_Generation_Tutorial.html
   Dataset_Generation_Tutorial_v6-0: Dataset_Generation_Tutorial_v6-0.html
-  Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
+  Dataset_Generation_Tutorial: Dataset_Generation_Tutorial.html
+  gcamreport_v6-0: gcamreport_v6-0.html
   Interactive_UI_Tutorial_v6-0: Interactive_UI_Tutorial_v6-0.html
+  Interactive_UI_Tutorial: Interactive_UI_Tutorial.html
   Modify_Mapping_Template_Tutorial: Modify_Mapping_Template_Tutorial.html
   README_v6-0: README_v6-0.html
-  Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
   Step_By_Step_Full_Example_v6-0: Step_By_Step_Full_Example_v6-0.html
+  Step_By_Step_Full_Example: Step_By_Step_Full_Example.html
   Version_Guide: Version_Guide.html
-  gcamreport_v6-0: gcamreport_v6-0.html
-last_built: 2024-04-08T07:52Z
+last_built: 2024-04-11T13:04Z
 urls:
   reference: https://bc3lc.github.io/gcamreport/reference
   article: https://bc3lc.github.io/gcamreport/articles
