+++ NEWS.md
@@ -1,8 +1,33 @@
+# gcamreport 7.0.1
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10960370.svg)](https://doi.org/10.5281/zenodo.10960370)
+
+`gcamreport` package suitable for GCAM 7.0 💥
+
+Novelties:
+
+- JOSS Publication.
+- Allow `gcamreport` to be used without cloning the repository or using the .Rproj file. Now suitable for running on clustered machines.
+- Allow to specify query files.
+- Add default options for selecting database or rgcam project scenarios.
+- Add tutorials for mapping modifications and query files specifications.
+- Add and fix user checks.
+- Improve and update package documentation.
+- Improve the package's robustness.
+- Optimize the package.
+- Clean up variables in cache.
+- Fix transport mapping.
+- Fix non-ASCII characters.
+- Reduce package dependencies.
+
+**Full Changelog**: https://github.com/bc3LC/gcamreport/compare/v7.0.0...v7.0.1
+
+
 # gcamreport 7.0.0
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10371722.svg)](https://doi.org/10.5281/zenodo.10371722)
 
-gcamreport package suitable for GCAM 7.0 💥
+`gcamreport` package suitable for GCAM 7.0 💥
 
 Novelties:
 
@@ -18,7 +43,7 @@ Novelties:
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8177345.svg)](https://doi.org/10.5281/zenodo.8177345)
 
-gcamreport package suitable for GCAM 6.0 💥
+`gcamreport` package suitable for GCAM 6.0 💥
 
 Novelties:
 
@@ -30,7 +55,7 @@ Novelties:
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7923878.svg)](https://doi.org/10.5281/zenodo.7923878)
 
-gcamreport package suitable for GCAM 6.0 with extra gas features 💥
+`gcamreport` package suitable for GCAM 6.0 with extra gas features 💥
 
 Novelties:
 
@@ -41,7 +66,7 @@ Novelties:
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7923791.svg)](https://doi.org/10.5281/zenodo.7923791)
 
-gcamreport package suitable for GCAM 6.0 💥
+`gcamreport` package suitable for GCAM 6.0 💥
 
 Novelties:
 
