### 2024-04-11 #5975 @klau506
Dear @martinfleis,

Thank you for all your work. I have checked all items on the post-review list and made a software release. The version number is [v.7.0.1](https://github.com/bc3LC/gcamreport/releases/tag/v7.0.1). It is available on Zenodo under this DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10960370.svg)](https://doi.org/10.5281/zenodo.10960370). The list of authors and their ORCIDS has been checked in the manuscript and in the Zenodo archive. The license has been checked.

Best,
Clàudia
