+++ man/iron_steel_trade_map.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{iron_steel_trade_map}
+\alias{iron_steel_trade_map}
+\title{iron_steel_trade_map}
+\format{
+.csv
+}
+\source{
+github
+}
+\usage{
+iron_steel_trade_map
+}
+\description{
+iron steel imports and exports map
+}
+\examples{
+\dontrun{
+library(gcamreport)
+gcamreport::iron_steel_trade_map
+}
+}
+\keyword{datasets}
