+++ man/get_iron_steel_exports.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/functions.R
+\name{get_iron_steel_exports}
+\alias{get_iron_steel_exports}
+\title{get_iron_steel_exports}
+\usage{
+get_iron_steel_exports()
+}
+\value{
+iron_steel_exports global variable
+}
+\description{
+Get iron steel production.
+}
+\keyword{industry}
+\keyword{internal}
+\keyword{tmp}
