+++ NAMESPACE
@@ -68,6 +68,9 @@ export(get_ghg_sector)
 export(get_gov_revenue)
 export(get_gov_revenue_sector)
 export(get_industry_production)
+export(get_iron_steel_clean)
+export(get_iron_steel_exports)
+export(get_iron_steel_imports)
 export(get_iron_steel_map)
 export(get_land)
 export(get_lu_co2)
@@ -131,6 +134,7 @@ importFrom(stats,complete.cases)
 importFrom(stringr,str_extract)
 importFrom(stringr,str_locate)
 importFrom(stringr,str_replace)
+importFrom(stringr,str_replace_all)
 importFrom(stringr,str_sub)
 importFrom(tibble,as_tibble)
 importFrom(tibble,tibble)
