### 2024-04-17 b2f4dcb
Fix desired regions filter when loading project
