### 2024-04-17 541c366
Fix desired regions filter when loading project
