+++ R/functions.R
@@ -693,7 +693,8 @@ get_ghg_sector <- function() {
     filter(variable %in% gcamreport::GHG_gases) %>%
     rename(ghg = variable) %>%
     left_join(filter_variables(gcamreport::kyoto_sector_map, "ghg_sector_clean"),
-              by = c('ghg', 'subsector', 'sector'), multiple = "all") %>%
+      by = c("ghg", "subsector", "sector"), multiple = "all"
+    ) %>%
     select(all_of(gcamreport::long_columns)) %>%
     bind_rows(
       LU_carbon_clean %>%
@@ -1228,6 +1229,7 @@ get_iron_steel_exports <- function() {
     left_join(filter_variables(gcamreport::iron_steel_trade_map, "iron_steel_clean"), by = c("sector")) %>%
     # extract region
     mutate(region = str_replace_all(subsector, " traded iron and steel", "")) %>%
+    filter(region %in% desired_regions) %>%
     # filter variables that are in terms of Mt
     group_by(scenario, region, var, year) %>%
     summarise(value = sum(value, na.rm = T)) %>%
@@ -1290,11 +1292,18 @@ get_ag_prices <- function() {
   ag_prices_clean <<-
     getQuery(prj, "prices by sector") %>%
     bind_rows(ag_prices_wld) %>%
-    left_join(filter_variables(gcamreport::ag_prices_map, "ag_prices_clean"), by = c("sector")) %>%
+    left_join(filter_variables(gcamreport::ag_prices_map, "ag_prices_clean"),
+      by = c("sector"), relationship = "many-to-many"
+    ) %>%
     filter(!is.na(var)) %>%
     group_by(scenario, region, sector) %>%
     mutate(value = value * unit_conv / value[year == 2005]) %>%
     ungroup() %>%
+    # do the mean by variable
+    group_by(scenario, region, var, year, ) %>%
+    summarise(value = mean(value)) %>%
+    ungroup() %>%
+    # rearrange dataset
     select(all_of(gcamreport::long_columns))
 }
 
