+++ NAMESPACE
@@ -19,6 +19,7 @@ export(do_data_sample)
 export(do_mount_tree)
 export(do_unmount_tree)
 export(filter_data_regions)
+export(filter_desired_regions)
 export(filter_loading_regions)
 export(filter_variables)
 export(gather_map)
