### 2024-04-20 #5975 @kthyng
Also please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
