+++ paper/paper.bib
@@ -14,7 +14,7 @@
     howpublished = {Online},
     month = jun,
     year = {2023},
-    url = {https://www.iamconsortium.org/event/iamc-webinar-standardization-of-scenario-data-collection-via-the-iamc-data-template-past-developments-and-future-challenges-volker-krey/},
+    url = {https://www.iamconsortium.org/event/iamc-webinar-standardization-of-scenario-data-collection-via-the-iamc- data-template-past-developments-and-future-challenges-volker-krey/},
 }
 
 @article{richters_ngfs_2022,
@@ -23,10 +23,12 @@
 	year = {2022},
 	note = {Publisher: Zenodo},
 	doi = {10.5281/zenodo.7085661},
+  publisher = {Zenodo},
+  url = {https://doi.org/10.5281/zenodo.7085661},
 }
 
 @article{masson-delmotte_climate_2021,
-	title = {Climate change 2021: the physical science basis},
+	title = {{Climate Change 2021: the physical science basis}},
 	volume = {2},
 	journal = {Contribution of working group I to the sixth assessment report of the intergovernmental panel on climate change},
 	author = {Masson-Delmotte, Valérie and Zhai, Panmao and Pirani, Anna and Connors, Sarah L and Péan, Clotilde and Berger, Sophie and Caud, Nada and Chen, Y and Goldfarb, L and Gomis, MI},
@@ -106,7 +108,7 @@
 
 
 @article{oneill2014,
-	title = {A new scenario framework for climate change research: the concept of shared socioeconomic pathways},
+	title = {A new scenario framework for climate change research: the concept of {Shared} {Socioeconomic} {Pathways}},
 	volume = {122},
 	issn = {0165-0009},
 	pages = {387--400},
@@ -144,7 +146,7 @@
 }
 
 @article{lin2020,
-	title = {The {TRUST} Principles for digital repositories},
+	title = {The {TRUST} {Principles} for digital repositories},
 	volume = {7},
 	issn = {2052-4463},
 	pages = {144},
@@ -158,12 +160,11 @@
 }
 
 @article{wilkinson2016,
-	title = {The {FAIR} Guiding Principles for scientific data management and stewardship},
+	title = {The {FAIR} {Guiding} {Principles} for scientific data management and stewardship},
 	volume = {3},
 	issn = {2052-4463},
 	url = {https://www.nature.com/articles/sdata201618},
 	doi = {10.1038/sdata.2016.18},
-	abstract = {In the realm of its Open Science strategy, the European Commission ({EC}) has announced its intentions to promote open access to research data resulting from publically funded research under Horizon 2020 to increase the circulation and exploitation of knowledge. Openness should become the default case respecting constraints to privacy, national security and intellectual property rights. For making open science happen, two major pillars need to be addressed. Data needs to be {FAIR} (Findable, Accessible, Interoperable, Reusable) to overcome the fragmentation and inefficiencies and thus to increase data reuse effectively. Commissioner Moedas made clear that "Europe's final transition must be one from fragmented data sets to an integrated European Open Science Cloud ({EOSC})". By 2020, all European researchers need to be able depositing, accessing and analyzing European scientific data through the {EOSC}. {EOSC} and {FAIR} research data are closely related and the {EC} is working out phases for implementing {EOSC} and action plans to make data {FAIR}.},
 	pages = {160018},
 	number = {1},
 	journaltitle = {Scientific data},
@@ -175,7 +176,7 @@
 }
 
 @article{shukla2022,
-	title = {Climate Change 2022 Mitigation of Climate Change},
[…]
