### 2024-04-21 #5975 @klau506
@kthyng Thank you so much for your checks. The Zenodo metadata title has been modified and all the capital letters in the bib file are now preserved.

Best,
Clàudia
